% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{And}
\alias{And}
\title{And}
\format{\code{\link{R6Class}} object.}
\usage{
And
}
\value{
Object of \code{\link{R6Class}} for modelling an And operator
}
\description{
And
}
\section{Methods}{

\describe{
 \item{\code{new(operations)}}{
   This method is used to instantiate an And operator
 }
}
}

\keyword{And}
\keyword{BinaryLogicOpType}
\keyword{Expression}
\keyword{OGC}
