% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{BinaryLogicOpType}
\alias{BinaryLogicOpType}
\title{BinaryLogicOpType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an BinaryLogicOpType
}
\description{
BinaryLogicOpType
}
\note{
abstract super class of all the binary logical operation classes
}
\section{Methods}{

\describe{
 \item{\code{new(...)}}{
   This method is used to instantiate an BinaryLogicOpType
 }
}
}

\keyword{BinaryLogicOpType}
\keyword{Expression}
\keyword{OGC}
