% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWDescribeRecord.R
\docType{class}
\name{CSWDescribeRecord}
\alias{CSWDescribeRecord}
\title{CSWDescribeRecord}
\format{\code{\link{R6Class}} object.}
\usage{
CSWDescribeRecord
}
\value{
Object of \code{\link{R6Class}} for modelling a CSW DescribeRecord request
}
\description{
CSWDescribeRecord
}
\section{Methods}{

\describe{
 \item{\code{new(url, version, namespace, ...)}}{
   This method is used to instantiate a CSWDescribeRecord object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CSW}
\keyword{DescribeRecord}
\keyword{OGC}
