% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWGetRecords.R
\docType{class}
\name{CSWGetRecords}
\alias{CSWGetRecords}
\title{CSWGetRecords}
\format{\code{\link{R6Class}} object.}
\usage{
CSWGetRecords
}
\value{
Object of \code{\link{R6Class}} for modelling a CSW GetRecords request
}
\description{
CSWGetRecords
}
\section{Methods}{

\describe{
 \item{\code{new(op, url, serviceVersion, user, pwd, query, logger, ...)}}{
   This method is used to instantiate a CSWGetRecords object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CSW}
\keyword{GetRecords}
\keyword{OGC}
