% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWTransaction.R
\docType{class}
\name{CSWTransaction}
\alias{CSWTransaction}
\title{CSWTransaction}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a CSW Transaction request
}
\description{
CSWTransaction
}
\note{
Class used internally by \pkg{ows4R} to trigger a CSW Transaction request
}
\section{Methods}{

\describe{
 \item{\code{new(op, url, serviceVersion, type, user, pwd, token, headers,
                 record, recordProperty, constraint, logger, ...)}}{
   This method is used to instantiate a CSWTransaction object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CSW}
\keyword{OGC}
\keyword{Transaction}
