% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSNamespace.R
\name{getOWSNamespace}
\alias{getOWSNamespace}
\title{getOWSNamespace}
\usage{
getOWSNamespace(id)
}
\arguments{
\item{id}{namespace prefix}
}
\description{
\code{getOWSNamespace} gets a namespace given its id
}
\examples{
            
  getOWSNamespace("GMD")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
