% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSNamespace.R
\name{getOWSNamespaces}
\alias{getOWSNamespaces}
\title{getOWSNamespaces}
\usage{
getOWSNamespaces()
}
\description{
\code{getOWSNamespaces} gets the list of namespaces registered
}
\examples{
            
  getOWSNamespaces()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
