% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{Not}
\alias{Not}
\title{Not}
\format{\code{\link{R6Class}} object.}
\usage{
Not
}
\value{
Object of \code{\link{R6Class}} for modelling an Not
}
\description{
Not
}
\section{Methods}{

\describe{
 \item{\code{new(...)}}{
   This method is used to instantiate an Not operator
 }
}
}

\examples{
  expr <- PropertyIsEqualTo$new(PropertyName = "property", Literal = "value")
  not <- Not$new(expr)
  not_xml <- not$encode() #see how it looks like in XML

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Expression}
\keyword{Not}
\keyword{OGC}
\keyword{UnaryLogicOpType}
