% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCAbstractObject.R
\docType{class}
\name{OGCAbstractObject}
\alias{OGCAbstractObject}
\title{OGCAbstractObject}
\format{\code{\link{R6Class}} object.}
\usage{
OGCAbstractObject
}
\value{
Object of \code{\link{R6Class}} for modelling an OGCAbstractObject
}
\description{
OGCAbstractObject
}
\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new()}}{
   This method is used to instantiate an OGCAbstractObject
 }
}
}

\keyword{Abstract}
\keyword{OGC}
\keyword{Object}
