% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{OGCExpression}
\alias{OGCExpression}
\title{OGCExpression}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Expression
}
\description{
OGCExpression
}
\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new(element, namespacePrefix, attrs, defaults, exprVersion)}}{
   This method is used to instantiate an \code{OGCExpression} object
 }
 \item{\code{setExprVersion(exprVersion)}}{
   Set expression version
 }
 \item{\code{getExprVersion()}}{
   Get expression version
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Expression}
\keyword{OGC}
