% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{Or}
\alias{Or}
\title{Or}
\format{\code{\link{R6Class}} object.}
\usage{
Or
}
\value{
Object of \code{\link{R6Class}} for modelling an Or operator
}
\description{
Or
}
\section{Methods}{

\describe{
 \item{\code{new(operations)}}{
   This method is used to instantiate an Or operator
 }
}
}

\keyword{BinaryLogicOpType}
\keyword{Expression}
\keyword{OGC}
\keyword{Or}
