% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSBoundingBox.R
\docType{class}
\name{OWSBoundingBox}
\alias{OWSBoundingBox}
\title{OWSBoundingBox}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Bounding Box
}
\description{
OWSBoundingBox
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, namespacePrefix, owsVersion, serviceVersion, logger)}}{
   This method is used to instantiate an \code{OWSBoundingBox} object
 }
 \item{\code{getBBOX()}}{
   Get BBOX as object of class \code{bbox} from \pkg{sf} package
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{OWS}
\keyword{boundingbox}
