% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSCapabilities.R
\docType{class}
\name{OWSCapabilities}
\alias{OWSCapabilities}
\title{OWSGetCapabilities}
\format{\code{\link{R6Class}} object.}
\usage{
OWSCapabilities
}
\value{
Object of \code{\link{R6Class}} with methods for interfacing an abstract
OWS Get Capabilities document.
}
\description{
OWSGetCapabilities
}
\section{Methods}{

\describe{
 \item{\code{new(url, service, serviceVersion, owsVersion, logger)}}{
   This method is used to instantiate a OWSGetCapabilities object
 }
 \item{\code{getUrl()}}{
   Get URL
 }
 \item{\code{getVersion()}}{
   Get version
 }
 \item{\code{getRequest()}}{
   Get request
 }
 \item{\code{getServiceIdentification()}}{
   Get the service identification
 }
 \item{\code{getServiceProvider()}}{
   Get the service provider
 }
 \item{\code{getOperationsMetadata()}}{
   Get the service operations metadata
 }
}
}

\examples{
\dontrun{
   OWSCapabilities$new("http://localhost:8080/geoserver/wfs", service = "wfs" version = "1.1.0")
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GetCapabilities}
\keyword{OGC}
\keyword{OWS}
