% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSCodeType.R
\docType{class}
\name{OWSCodeType}
\alias{OWSCodeType}
\title{OWSCodeType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OWS CodeType
}
\description{
OWSCodeType
}
\section{Methods}{

\describe{
 \item{\code{new(expr)}}{
   This method is used to instantiate an OWSCodeType object. The unique
   argument should be an object of class \code{character}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CodeType}
\keyword{OWS}
