% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSHttpRequest.R
\docType{class}
\name{OWSHttpRequest}
\alias{OWSHttpRequest}
\title{OWSHttpRequest}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a generic OWS http request
}
\description{
OWSHttpRequest

OWSHttpRequest
}
\note{
Abstract class used internally by \pkg{ows4R}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{HTTP}
\keyword{OGC}
\keyword{OWS}
\keyword{Request}
\section{Super class}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{OWSHttpRequest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OWSHttpRequest$new()}}
\item \href{#method-execute}{\code{OWSHttpRequest$execute()}}
\item \href{#method-getCapabilities}{\code{OWSHttpRequest$getCapabilities()}}
\item \href{#method-getRequest}{\code{OWSHttpRequest$getRequest()}}
\item \href{#method-getRequestHeaders}{\code{OWSHttpRequest$getRequestHeaders()}}
\item \href{#method-getStatus}{\code{OWSHttpRequest$getStatus()}}
\item \href{#method-getResponse}{\code{OWSHttpRequest$getResponse()}}
\item \href{#method-getException}{\code{OWSHttpRequest$getException()}}
\item \href{#method-getResult}{\code{OWSHttpRequest$getResult()}}
\item \href{#method-setResult}{\code{OWSHttpRequest$setResult()}}
\item \href{#method-clone}{\code{OWSHttpRequest$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode">}\href{../../ows4R/html/OGCAbstractObject.html#method-encode}{\code{ows4R::OGCAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR">}\href{../../ows4R/html/OGCAbstractObject.html#method-ERROR}{\code{ows4R::OGCAbstractObject$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClass}{\code{ows4R::OGCAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClassName}{\code{ows4R::OGCAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition">}\href{../../ows4R/html/OGCAbstractObject.html#method-getNamespaceDefinition}{\code{ows4R::OGCAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO">}\href{../../ows4R/html/OGCAbstractObject.html#method-INFO}{\code{ows4R::OGCAbstractObject$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom">}\href{../../ows4R/html/OGCAbstractObject.html#method-isFieldInheritedFrom}{\code{ows4R::OGCAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger">}\href{../../ows4R/html/OGCAbstractObject.html#method-logger}{\code{ows4R::OGCAbstractObject$logger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN">}\href{../../ows4R/html/OGCAbstractObject.html#method-WARN}{\code{ows4R::OGCAbstractObject$WARN()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes an OWS HTTP request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$new(
  element,
  namespacePrefix,
  capabilities,
  op,
  type,
  url,
  request,
  user = NULL,
  pwd = NULL,
  token = NULL,
  headers = c(),
  namedParams = NULL,
  attrs = NULL,
  contentType = "text/xml",
  mimeType = "text/xml",
  logger = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{element}}{element}

\item{\code{namespacePrefix}}{namespace prefix}

\item{\code{capabilities}}{object of class or extending \link{OWSCapabilities}}

\item{\code{op}}{object of class \link{OWSOperation}}

\item{\code{type}}{type of request, eg. GET, POST}

\item{\code{url}}{url}

\item{\code{request}}{request name}

\item{\code{user}}{user}

\item{\code{pwd}}{password}

\item{\code{token}}{token}

\item{\code{headers}}{headers}

\item{\code{namedParams}}{a named \code{list}}

\item{\code{attrs}}{attributes}

\item{\code{contentType}}{content type. Default value is "text/xml"}

\item{\code{mimeType}}{mime type. Default value is "text/xml"}

\item{\code{logger}}{logger}

\item{\code{...}}{any other parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
Executes the request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$execute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCapabilities"></a>}}
\if{latex}{\out{\hypertarget{method-getCapabilities}{}}}
\subsection{Method \code{getCapabilities()}}{
Get capabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$getCapabilities()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class or extending \link{OWSCapabilities}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRequest"></a>}}
\if{latex}{\out{\hypertarget{method-getRequest}{}}}
\subsection{Method \code{getRequest()}}{
Get request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$getRequest()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRequestHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-getRequestHeaders}{}}}
\subsection{Method \code{getRequestHeaders()}}{
Get request headers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$getRequestHeaders()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the request headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStatus"></a>}}
\if{latex}{\out{\hypertarget{method-getStatus}{}}}
\subsection{Method \code{getStatus()}}{
get status code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$getStatus()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the request status code
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getResponse"></a>}}
\if{latex}{\out{\hypertarget{method-getResponse}{}}}
\subsection{Method \code{getResponse()}}{
get request response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$getResponse()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the request response
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getException"></a>}}
\if{latex}{\out{\hypertarget{method-getException}{}}}
\subsection{Method \code{getException()}}{
get request exception
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$getException()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the request exception
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getResult"></a>}}
\if{latex}{\out{\hypertarget{method-getResult}{}}}
\subsection{Method \code{getResult()}}{
Get the result \code{TRUE} if the request is successful, \code{FALSE} otherwise
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$getResult()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the result, object of class \code{logical}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setResult"></a>}}
\if{latex}{\out{\hypertarget{method-setResult}{}}}
\subsection{Method \code{setResult()}}{
Set the result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$setResult(result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{object of class \code{logical}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSHttpRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
