% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSLogger.R
\docType{class}
\name{OWSLogger}
\alias{OWSLogger}
\title{OWSLogger}
\format{\code{\link{R6Class}} object.}
\usage{
OWSLogger
}
\value{
Object of \code{\link{R6Class}} to provide logs to \pkg{ows4R}
}
\description{
OWSLogger
}
\section{Fields}{

\describe{
\item{\code{loggerType}}{the type of logger}

\item{\code{verbose.info}}{if basic logs have to be printed}

\item{\code{verbose.debug}}{if verbose logs have to be printed}
}}

\note{
Internal class only.
}
\section{Methods}{

\describe{
 \item{\code{logger(type, text)}}{
   Basic logger to report logs. Used internally
 }
 \item{\code{INFO(text)}}{
   Logger to report information. Used internally
 }
 \item{\code{WARN(text)}}{
   Logger to report warnings. Used internally
 }
 \item{\code{ERROR(text)}}{
   Logger to report errors. Used internally
 }
 \item{\code{getClassName()}}{
   Retrieves the name of the class instance
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{datasets}
