% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSNamespace.R
\docType{class}
\name{OWSNamespace}
\alias{OWSNamespace}
\title{OWSNamespace}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OWS Namespace
}
\description{
OWSNamespace
}
\note{
class used internally by ows4R for specifying XML namespaces
}
\section{Methods}{

\describe{
 \item{\code{new(id, uri)}}{
   This method is used to instantiate an OWS namespace
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OWS}
\keyword{namespace}
