% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSOperationsMetadata.R
\docType{class}
\name{OWSOperationsMetadata}
\alias{OWSOperationsMetadata}
\title{OWSOperationsMetadata}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Operations Metadata
}
\description{
OWSOperationsMetadata

OWSOperationsMetadata
}
\note{
Abstract class used internally by \pkg{ows4R}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{OWS}
\keyword{metadata}
\keyword{operation}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OWSOperationsMetadata-new}{\code{OWSOperationsMetadata$new()}}
\item \href{#method-OWSOperationsMetadata-getOperations}{\code{OWSOperationsMetadata$getOperations()}}
\item \href{#method-OWSOperationsMetadata-clone}{\code{OWSOperationsMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OWSOperationsMetadata-new"></a>}}
\if{latex}{\out{\hypertarget{method-OWSOperationsMetadata-new}{}}}
\subsection{Method \code{new()}}{
Initializes an \link{OWSOperationsMetadata} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSOperationsMetadata$new(xmlObj, owsVersion, serviceVersion)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmlObj}}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{\code{owsVersion}}{OWS version}

\item{\code{serviceVersion}}{service version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OWSOperationsMetadata-getOperations"></a>}}
\if{latex}{\out{\hypertarget{method-OWSOperationsMetadata-getOperations}{}}}
\subsection{Method \code{getOperations()}}{
Get operations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSOperationsMetadata$getOperations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list of \link{OWSOperation}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OWSOperationsMetadata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OWSOperationsMetadata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSOperationsMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
