% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSServiceIdentification.R
\docType{class}
\name{OWSServiceIdentification}
\alias{OWSServiceIdentification}
\title{OWSServiceIdentification}
\format{\code{\link{R6Class}} object.}
\usage{
OWSServiceIdentification
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Service Identification
}
\description{
OWSServiceIdentification
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, serviceVersion)}}{
   This method is used to instantiate a OWSServiceIdentification object
 }
 \item{\code{getName()}}{
   Get service name
 }
 \item{\code{getTitle()}}{
   Get service title
 }
 \item{\code{getAbstract()}}{
   Get service abstract
 }
 \item{\code{getKeywords()}}{
   Get service keywords
 }
 \item{\code{getOnlineResource()}}{
   Get service online resource
 }
 \item{\code{getServiceType}}{
   Get service type
 }
 \item{\code{getServiceTypeVersion}}{
  Get service type version
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Identification}
\keyword{OGC}
\keyword{OWS}
\keyword{Service}
