% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSUtils.R
\docType{data}
\name{OWSUtils}
\alias{OWSUtils}
\title{OWSUtils}
\format{\code{\link{R6Class}} object.}
\usage{
OWSUtils
}
\value{
set of OWS Utilities
}
\description{
OWSUtils
}
\section{Static methods}{

\describe{
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{OWS}
\keyword{Utils}
