% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsBetween}
\alias{PropertyIsBetween}
\title{PropertyIsBetween}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsBetween
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsBetween
}
\description{
PropertyIsBetween
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName, lower, upper)}}{
   This method is used to instantiate an PropertyIsBetween
 }
}
}

\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsBetween}
