% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsEqualTo}
\alias{PropertyIsEqualTo}
\title{PropertyIsEqualTo}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsEqualTo
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsEqualTo
}
\description{
PropertyIsEqualTo
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName, Literal, matchCase)}}{
   This method is used to instantiate an PropertyIsEqualTo
 }
}
}

\keyword{BinaryComparisonOpType}
\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsEqualTo}
