% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsGreaterThan}
\alias{PropertyIsGreaterThan}
\title{PropertyIsGreaterThan}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsGreaterThan
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsGreaterThan
}
\description{
PropertyIsGreaterThan
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName, Literal, matchCase)}}{
   This method is used to instantiate an PropertyIsGreaterThan
 }
}
}

\keyword{BinaryComparisonOpType}
\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsGreaterThan}
