% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsLessThanOrEqualTo}
\alias{PropertyIsLessThanOrEqualTo}
\title{PropertyIsLesserThanOrEqualTo}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsLessThanOrEqualTo
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsLesserThanOrEqualTo
}
\description{
PropertyIsLesserThanOrEqualTo
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName, Literal, matchCase)}}{
   This method is used to instantiate an PropertyIsLesserThanOrEqualTo
 }
}
}

\examples{
  expr <- PropertyIsLessThanOrEqualTo$new(PropertyName = "property", Literal = "value")
  expr_xml <- expr$encode() #see how it looks like in XML
  
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{BinaryComparisonOpType}
\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsLesserThanOrEqualTo}
