% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCSCoverage.R
\docType{class}
\name{WCSCoverage}
\alias{WCSCoverage}
\title{WCSCoverage}
\format{
\code{\link[R6]{R6Class}} object.
}
\value{
Object of \code{\link[R6]{R6Class}} modelling a WCS coverage
}
\description{
WCSCoverage

WCSCoverage
}
\note{
Class used internally by ows4R.
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Coverage}
\keyword{OGC}
\keyword{WCS}
\section{Super class}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{WCSCoverage}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{description}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WCSCoverage-new}{\code{WCSCoverage$new()}}
\item \href{#method-WCSCoverage-getIdentifier}{\code{WCSCoverage$getIdentifier()}}
\item \href{#method-WCSCoverage-getTitle}{\code{WCSCoverage$getTitle()}}
\item \href{#method-WCSCoverage-getAbstract}{\code{WCSCoverage$getAbstract()}}
\item \href{#method-WCSCoverage-getReference}{\code{WCSCoverage$getReference()}}
\item \href{#method-WCSCoverage-getData}{\code{WCSCoverage$getData()}}
\item \href{#method-WCSCoverage-clone}{\code{WCSCoverage$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-ERROR'><code>ows4R::OGCAbstractObject$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-INFO'><code>ows4R::OGCAbstractObject$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-WARN'><code>ows4R::OGCAbstractObject$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-encode'><code>ows4R::OGCAbstractObject$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-getClass'><code>ows4R::OGCAbstractObject$getClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-getClassName'><code>ows4R::OGCAbstractObject$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-getNamespaceDefinition'><code>ows4R::OGCAbstractObject$getNamespaceDefinition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-isFieldInheritedFrom'><code>ows4R::OGCAbstractObject$isFieldInheritedFrom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-logger'><code>ows4R::OGCAbstractObject$logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="print"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-print'><code>ows4R::OGCAbstractObject$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverage-new"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverage-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{WCSCoverage}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverage$new(xmlObj, serviceVersion, owsVersion, logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmlObj}}{object of class \link[XML]{XMLInternalNode-class}}

\item{\code{serviceVersion}}{WCS service version}

\item{\code{owsVersion}}{OWS version}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverage-getIdentifier"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverage-getIdentifier}{}}}
\subsection{Method \code{getIdentifier()}}{
Get identifier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverage$getIdentifier()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverage-getTitle"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverage-getTitle}{}}}
\subsection{Method \code{getTitle()}}{
Get title
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverage$getTitle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverage-getAbstract"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverage-getAbstract}{}}}
\subsection{Method \code{getAbstract()}}{
Get abstract
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverage$getAbstract()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverage-getReference"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverage-getReference}{}}}
\subsection{Method \code{getReference()}}{
Get reference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverage$getReference()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverage-getData"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverage-getData}{}}}
\subsection{Method \code{getData()}}{
Get data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverage$getData(filename = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename. Optional file name where to download the coverage}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class \code{SpatRaster} from \pkg{terra}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
