% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCSCoverageDescription.R
\docType{class}
\name{WCSCoverageDescription}
\alias{WCSCoverageDescription}
\title{WCSCoverageDescription}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WCS coverage summary
}
\description{
WCSCoverageDescription

WCSCoverageDescription
}
\note{
Class used internally by ows4R.
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Coverage}
\keyword{OGC}
\keyword{WCS}
\section{Super classes}{
\code{\link[geometa:geometaLogger]{geometa::geometaLogger}} -> \code{\link[geometa:ISOAbstractObject]{geometa::ISOAbstractObject}} -> \code{\link[geometa:GMLAbstractObject]{geometa::GMLAbstractObject}} -> \code{\link[geometa:GMLAbstractFeature]{geometa::GMLAbstractFeature}} -> \code{\link[geometa:GMLAbstractCoverage]{geometa::GMLAbstractCoverage}} -> \code{\link[geometa:GMLCOVAbstractCoverage]{geometa::GMLCOVAbstractCoverage}} -> \code{WCSCoverageDescription}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CoverageId}}{coverage ID}

\item{\code{SupportedCRS}}{supported CRS}

\item{\code{SupportedFormat}}{supported Format}

\item{\code{Domain}}{domain}

\item{\code{Range}}{range}

\item{\code{ServiceParameters}}{service parmaeters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WCSCoverageDescription$new()}}
\item \href{#method-getId}{\code{WCSCoverageDescription$getId()}}
\item \href{#method-getSupportedCRS}{\code{WCSCoverageDescription$getSupportedCRS()}}
\item \href{#method-getSupportedFormats}{\code{WCSCoverageDescription$getSupportedFormats()}}
\item \href{#method-getDomain}{\code{WCSCoverageDescription$getDomain()}}
\item \href{#method-getRange}{\code{WCSCoverageDescription$getRange()}}
\item \href{#method-clone}{\code{WCSCoverageDescription$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="ERROR">}\href{../../geometa/html/geometaLogger.html#method-ERROR}{\code{geometa::geometaLogger$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="INFO">}\href{../../geometa/html/geometaLogger.html#method-INFO}{\code{geometa::geometaLogger$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="WARN">}\href{../../geometa/html/geometaLogger.html#method-WARN}{\code{geometa::geometaLogger$WARN()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addFieldAttrs">}\href{../../geometa/html/ISOAbstractObject.html#method-addFieldAttrs}{\code{geometa::ISOAbstractObject$addFieldAttrs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addListElement">}\href{../../geometa/html/ISOAbstractObject.html#method-addListElement}{\code{geometa::ISOAbstractObject$addListElement()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="contains">}\href{../../geometa/html/ISOAbstractObject.html#method-contains}{\code{geometa::ISOAbstractObject$contains()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="createLocalisedProperty">}\href{../../geometa/html/ISOAbstractObject.html#method-createLocalisedProperty}{\code{geometa::ISOAbstractObject$createLocalisedProperty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="decode">}\href{../../geometa/html/ISOAbstractObject.html#method-decode}{\code{geometa::ISOAbstractObject$decode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="delListElement">}\href{../../geometa/html/ISOAbstractObject.html#method-delListElement}{\code{geometa::ISOAbstractObject$delListElement()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="encode">}\href{../../geometa/html/ISOAbstractObject.html#method-encode}{\code{geometa::ISOAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClass">}\href{../../geometa/html/ISOAbstractObject.html#method-getClass}{\code{geometa::ISOAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClassName">}\href{../../geometa/html/ISOAbstractObject.html#method-getClassName}{\code{geometa::ISOAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getNamespaceDefinition">}\href{../../geometa/html/ISOAbstractObject.html#method-getNamespaceDefinition}{\code{geometa::ISOAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isDocument">}\href{../../geometa/html/ISOAbstractObject.html#method-isDocument}{\code{geometa::ISOAbstractObject$isDocument()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isFieldInheritedFrom">}\href{../../geometa/html/ISOAbstractObject.html#method-isFieldInheritedFrom}{\code{geometa::ISOAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="print">}\href{../../geometa/html/ISOAbstractObject.html#method-print}{\code{geometa::ISOAbstractObject$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="save">}\href{../../geometa/html/ISOAbstractObject.html#method-save}{\code{geometa::ISOAbstractObject$save()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setAttr">}\href{../../geometa/html/ISOAbstractObject.html#method-setAttr}{\code{geometa::ISOAbstractObject$setAttr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeList">}\href{../../geometa/html/ISOAbstractObject.html#method-setCodeList}{\code{geometa::ISOAbstractObject$setCodeList()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeListValue">}\href{../../geometa/html/ISOAbstractObject.html#method-setCodeListValue}{\code{geometa::ISOAbstractObject$setCodeListValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeSpace">}\href{../../geometa/html/ISOAbstractObject.html#method-setCodeSpace}{\code{geometa::ISOAbstractObject$setCodeSpace()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setHref">}\href{../../geometa/html/ISOAbstractObject.html#method-setHref}{\code{geometa::ISOAbstractObject$setHref()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setId">}\href{../../geometa/html/ISOAbstractObject.html#method-setId}{\code{geometa::ISOAbstractObject$setId()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setIsNull">}\href{../../geometa/html/ISOAbstractObject.html#method-setIsNull}{\code{geometa::ISOAbstractObject$setIsNull()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setValue">}\href{../../geometa/html/ISOAbstractObject.html#method-setValue}{\code{geometa::ISOAbstractObject$setValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="validate">}\href{../../geometa/html/ISOAbstractObject.html#method-validate}{\code{geometa::ISOAbstractObject$validate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="wrapBaseElement">}\href{../../geometa/html/ISOAbstractObject.html#method-wrapBaseElement}{\code{geometa::ISOAbstractObject$wrapBaseElement()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="GMLAbstractFeature" data-id="setBoundedBy">}\href{../../geometa/html/GMLAbstractFeature.html#method-setBoundedBy}{\code{geometa::GMLAbstractFeature$setBoundedBy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="GMLAbstractCoverage" data-id="setDomainSet">}\href{../../geometa/html/GMLAbstractCoverage.html#method-setDomainSet}{\code{geometa::GMLAbstractCoverage$setDomainSet()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="GMLAbstractCoverage" data-id="setRangeSet">}\href{../../geometa/html/GMLAbstractCoverage.html#method-setRangeSet}{\code{geometa::GMLAbstractCoverage$setRangeSet()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{WCSCoverageDescription}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDescription$new(xmlObj, serviceVersion, owsVersion, logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmlObj}}{an object of class \link{XMLInternalNode-class} to initialize from XML}

\item{\code{serviceVersion}}{service version}

\item{\code{owsVersion}}{OWS version}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getId"></a>}}
\if{latex}{\out{\hypertarget{method-getId}{}}}
\subsection{Method \code{getId()}}{
getId
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDescription$getId()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the coverage id, object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSupportedCRS"></a>}}
\if{latex}{\out{\hypertarget{method-getSupportedCRS}{}}}
\subsection{Method \code{getSupportedCRS()}}{
getSupported CRS. Applies to WCS 1 coverage descriptions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDescription$getSupportedCRS()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSupportedFormats"></a>}}
\if{latex}{\out{\hypertarget{method-getSupportedFormats}{}}}
\subsection{Method \code{getSupportedFormats()}}{
get supported formats. Applies to WCS 1 coverage descriptions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDescription$getSupportedFormats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDomain"></a>}}
\if{latex}{\out{\hypertarget{method-getDomain}{}}}
\subsection{Method \code{getDomain()}}{
get domain. Applies to WCS 1 coverage descriptions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDescription$getDomain()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRange"></a>}}
\if{latex}{\out{\hypertarget{method-getRange}{}}}
\subsection{Method \code{getRange()}}{
get range. Applies to WCS 1.0 coverage descriptions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDescription$getRange()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDescription$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
