% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCSCoverageFilenameHandler.R
\name{WCSCoverageFilenameHandler}
\alias{WCSCoverageFilenameHandler}
\title{WCSCoverageFilenameHandler}
\usage{
WCSCoverageFilenameHandler(identifier, time, elevation, bbox, format)
}
\arguments{
\item{identifier}{coverage identifier}

\item{time}{time}

\item{elevation}{elevation}

\item{bbox}{bbox}

\item{format}{format}
}
\value{
the filename to use for coverage download
}
\description{
\code{WCSCoverageFilenameHandler} provides a coverage filename handler for coverage download
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
