% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCSCoverageSpatialDomain.R
\docType{class}
\name{WCSCoverageSpatialDomain}
\alias{WCSCoverageSpatialDomain}
\title{WCSCoverageSpatialDomain}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WCS coverage spatial domain
}
\description{
WCSCoverageSpatialDomain

WCSCoverageSpatialDomain
}
\note{
Class used internally by ows4R.
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Coverage}
\keyword{OGC}
\keyword{WCS}
\keyword{domain}
\keyword{spatial}
\section{Super class}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{WCSCoverageSpatialDomain}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{envelopes}}{envelopes. For WCS 1.0}

\item{\code{BoundingBox}}{bounding box. For WCS 1.1}

\item{\code{grids}}{For WCS 1.0}

\item{\code{GridCRS}}{grid CRS. For WCS 1.1}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WCSCoverageSpatialDomain$new()}}
\item \href{#method-getEnvelopes}{\code{WCSCoverageSpatialDomain$getEnvelopes()}}
\item \href{#method-getBoundingBox}{\code{WCSCoverageSpatialDomain$getBoundingBox()}}
\item \href{#method-getGrids}{\code{WCSCoverageSpatialDomain$getGrids()}}
\item \href{#method-getGridCRS}{\code{WCSCoverageSpatialDomain$getGridCRS()}}
\item \href{#method-clone}{\code{WCSCoverageSpatialDomain$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode">}\href{../../ows4R/html/OGCAbstractObject.html#method-encode}{\code{ows4R::OGCAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR">}\href{../../ows4R/html/OGCAbstractObject.html#method-ERROR}{\code{ows4R::OGCAbstractObject$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClass}{\code{ows4R::OGCAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClassName}{\code{ows4R::OGCAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition">}\href{../../ows4R/html/OGCAbstractObject.html#method-getNamespaceDefinition}{\code{ows4R::OGCAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO">}\href{../../ows4R/html/OGCAbstractObject.html#method-INFO}{\code{ows4R::OGCAbstractObject$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom">}\href{../../ows4R/html/OGCAbstractObject.html#method-isFieldInheritedFrom}{\code{ows4R::OGCAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger">}\href{../../ows4R/html/OGCAbstractObject.html#method-logger}{\code{ows4R::OGCAbstractObject$logger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN">}\href{../../ows4R/html/OGCAbstractObject.html#method-WARN}{\code{ows4R::OGCAbstractObject$WARN()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{WCSCoverageDomain}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSpatialDomain$new(xmlObj, serviceVersion, owsVersion, logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmlObj}}{an object of class \link{XMLInternalNode-class} to initialize from XML}

\item{\code{serviceVersion}}{service version}

\item{\code{owsVersion}}{OWS version}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEnvelopes"></a>}}
\if{latex}{\out{\hypertarget{method-getEnvelopes}{}}}
\subsection{Method \code{getEnvelopes()}}{
Get envelopes. Method that applies to WCS 1.0 only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSpatialDomain$getEnvelopes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list of objects of class \link{GMLEnvelope} or \link{GMLEnvelopeWithTimePeriod}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBoundingBox"></a>}}
\if{latex}{\out{\hypertarget{method-getBoundingBox}{}}}
\subsection{Method \code{getBoundingBox()}}{
Get bounding boxes. Method that applies to WCS 1.1 only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSpatialDomain$getBoundingBox()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list of objects of class \link{OWSBoundingBox}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGrids"></a>}}
\if{latex}{\out{\hypertarget{method-getGrids}{}}}
\subsection{Method \code{getGrids()}}{
Get grids. Method that applies to WCS 1.0 only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSpatialDomain$getGrids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list of of objects of class \link{GMLGrid} or \link{GMLRectifiedGrid}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGridCRS"></a>}}
\if{latex}{\out{\hypertarget{method-getGridCRS}{}}}
\subsection{Method \code{getGridCRS()}}{
Get Grid CRS. Method that applies to WCS 1.1 only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSpatialDomain$getGridCRS()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list of objects of class \link{WCSGridCRS}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSpatialDomain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
