% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCSCoverageSummary.R
\docType{class}
\name{WCSCoverageSummary}
\alias{WCSCoverageSummary}
\title{WCSCoverageSummary}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WCS coverage summary
}
\description{
WCSCoverageSummary

WCSCoverageSummary
}
\note{
Class used internally by ows4R.
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, capabilities, serviceVersion, owsVersion, logger)}}{
   This method is used to instantiate a \code{WCSCoverageSummary} object
 }
 \item{\code{getId()}}{
   Get the identifier
 }
 \item{\code{getSubtype()}}{
   Get the coverage subtype
 }
 \item{\code{getSubtypeParent()}}{
   Get the coverage subtype parent type(s)
 }
 \item{\code{getWGS84BoundingBox()}}{
   Get coverage WGS84 bounding box as list of \code{OWSWGS84BoundingBox}
 }
 \item{\code{getBoundingBox()}}{
   Get coverage bounding box as list of \code{OWSBoundingBox}
 }
 \item{\code{getDescription()}}{
   Get coverage description
 }
 \item{\code{getDimensions()}}{
   Get the coverage dimensions, returned as\code{list}. Each element of this list
   will reference the axis abbreviation/label used for coverage query based on this
   dimension, the \pkg{geometa} GML object and classname, the type of dimension such as
   'geographic', 'temporal' or 'vertical', and if applicable: the possible values 
   (coefficients) that can be used for querying the coverage on the dimension. 
 }
 \item{\code{getCoverage(bbox, crs, time, elevation, format, rangesubset, 
                         gridbaseCRS, gridtype, gridCS, gridorigin, gridoffsets)}}{
   Get the coverage data as object of class \code{RasterLayer} from \pkg{raster}
 }
 \item{\code{getCoverageStack(time, elevation, bbox)}}{
   Get a stack of coverages (also known as 'datacube') for time and/or elevation. Each parameter
   can be specified a vector of values (coefficients) as filter. If no value is specified, the function will
   consider all coefficients for the given dimension. An additional bbox parameter allows to restrain
   the geographic extent of the overall coverage stack/datacube.
   
   For a spatio-temporal 3D coverage, in case that no time filter is specified, the full temporal coverage stack
   will be downloaded. For a spatio/vertical 3D coverage, in case that no elevation filter is specified, the full
   vertical coverage stack will be downloaded. Finally, in case of spatio-vertical-temporal 4D coverage, in case
   no time/elevation filter is specified, the full coverage stack corressponding to the combinations of temporal
   and vertical coefficients will be downloaded.
   
   This function returns an object of class \code{RasterStack} from \pkg{raster}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Coverage}
\keyword{OGC}
\keyword{WCS}
\section{Super class}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{WCSCoverageSummary}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CoverageId}}{coverage id}

\item{\code{CoverageSubtype}}{coverage subtype}

\item{\code{CoverageSubtypeParent}}{coverage subtype parent}

\item{\code{WGS84BoundingBox}}{WGS84 bounding box}

\item{\code{BoundingBox}}{bounding box}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WCSCoverageSummary$new()}}
\item \href{#method-getId}{\code{WCSCoverageSummary$getId()}}
\item \href{#method-getSubtype}{\code{WCSCoverageSummary$getSubtype()}}
\item \href{#method-getSubtypeParent}{\code{WCSCoverageSummary$getSubtypeParent()}}
\item \href{#method-getWGS84BoundingBox}{\code{WCSCoverageSummary$getWGS84BoundingBox()}}
\item \href{#method-getBoundingBox}{\code{WCSCoverageSummary$getBoundingBox()}}
\item \href{#method-getDescription}{\code{WCSCoverageSummary$getDescription()}}
\item \href{#method-clone}{\code{WCSCoverageSummary$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode">}\href{../../ows4R/html/OGCAbstractObject.html#method-encode}{\code{ows4R::OGCAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR">}\href{../../ows4R/html/OGCAbstractObject.html#method-ERROR}{\code{ows4R::OGCAbstractObject$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClass}{\code{ows4R::OGCAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClassName}{\code{ows4R::OGCAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition">}\href{../../ows4R/html/OGCAbstractObject.html#method-getNamespaceDefinition}{\code{ows4R::OGCAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO">}\href{../../ows4R/html/OGCAbstractObject.html#method-INFO}{\code{ows4R::OGCAbstractObject$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom">}\href{../../ows4R/html/OGCAbstractObject.html#method-isFieldInheritedFrom}{\code{ows4R::OGCAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger">}\href{../../ows4R/html/OGCAbstractObject.html#method-logger}{\code{ows4R::OGCAbstractObject$logger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN">}\href{../../ows4R/html/OGCAbstractObject.html#method-WARN}{\code{ows4R::OGCAbstractObject$WARN()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{WCSCoverageSummary} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$new(
  xmlObj,
  capabilities,
  serviceVersion,
  owsVersion,
  logger = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmlObj}}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{\code{capabilities}}{object of class \link{WCSCapabilities}}

\item{\code{serviceVersion}}{WCS service version}

\item{\code{owsVersion}}{version}

\item{\code{logger}}{logger type \code{NULL}, "INFO" or "DEBUG"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getId"></a>}}
\if{latex}{\out{\hypertarget{method-getId}{}}}
\subsection{Method \code{getId()}}{
Get coverage ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getId()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSubtype"></a>}}
\if{latex}{\out{\hypertarget{method-getSubtype}{}}}
\subsection{Method \code{getSubtype()}}{
Get sub type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getSubtype()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSubtypeParent"></a>}}
\if{latex}{\out{\hypertarget{method-getSubtypeParent}{}}}
\subsection{Method \code{getSubtypeParent()}}{
Get sub type parent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getSubtypeParent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWGS84BoundingBox"></a>}}
\if{latex}{\out{\hypertarget{method-getWGS84BoundingBox}{}}}
\subsection{Method \code{getWGS84BoundingBox()}}{
Get bounding box
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getWGS84BoundingBox()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{OWSWGS84BoundingBox}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBoundingBox"></a>}}
\if{latex}{\out{\hypertarget{method-getBoundingBox}{}}}
\subsection{Method \code{getBoundingBox()}}{
Get WGS84 bounding box
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getBoundingBox()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{OWSBoundingBox}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDescription"></a>}}
\if{latex}{\out{\hypertarget{method-getDescription}{}}}
\subsection{Method \code{getDescription()}}{
Get description
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getDescription()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{WCSCoverageDescription}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
