% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCSGSElevationDomain.R
\docType{class}
\name{WCSGSElevationDomain}
\alias{WCSGSElevationDomain}
\title{WCSGSElevationDomain}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WCS geoserver elevation domain object
}
\description{
WCSGSElevationDomain

WCSGSElevationDomain
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GeoServer}
\keyword{WCS}
\keyword{domain}
\keyword{elevation}
\section{Super classes}{
\code{\link[geometa:geometaLogger]{geometa::geometaLogger}} -> \code{\link[geometa:ISOAbstractObject]{geometa::ISOAbstractObject}} -> \code{WCSGSElevationDomain}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{SingleValue}}{single values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-decode}{\code{WCSGSElevationDomain$decode()}}
\item \href{#method-clone}{\code{WCSGSElevationDomain$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="ERROR">}\href{../../geometa/html/geometaLogger.html#method-ERROR}{\code{geometa::geometaLogger$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="INFO">}\href{../../geometa/html/geometaLogger.html#method-INFO}{\code{geometa::geometaLogger$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="WARN">}\href{../../geometa/html/geometaLogger.html#method-WARN}{\code{geometa::geometaLogger$WARN()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addFieldAttrs">}\href{../../geometa/html/ISOAbstractObject.html#method-addFieldAttrs}{\code{geometa::ISOAbstractObject$addFieldAttrs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addListElement">}\href{../../geometa/html/ISOAbstractObject.html#method-addListElement}{\code{geometa::ISOAbstractObject$addListElement()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="contains">}\href{../../geometa/html/ISOAbstractObject.html#method-contains}{\code{geometa::ISOAbstractObject$contains()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="createLocalisedProperty">}\href{../../geometa/html/ISOAbstractObject.html#method-createLocalisedProperty}{\code{geometa::ISOAbstractObject$createLocalisedProperty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="delListElement">}\href{../../geometa/html/ISOAbstractObject.html#method-delListElement}{\code{geometa::ISOAbstractObject$delListElement()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="encode">}\href{../../geometa/html/ISOAbstractObject.html#method-encode}{\code{geometa::ISOAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClass">}\href{../../geometa/html/ISOAbstractObject.html#method-getClass}{\code{geometa::ISOAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClassName">}\href{../../geometa/html/ISOAbstractObject.html#method-getClassName}{\code{geometa::ISOAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getNamespaceDefinition">}\href{../../geometa/html/ISOAbstractObject.html#method-getNamespaceDefinition}{\code{geometa::ISOAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="initialize">}\href{../../geometa/html/ISOAbstractObject.html#method-initialize}{\code{geometa::ISOAbstractObject$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isDocument">}\href{../../geometa/html/ISOAbstractObject.html#method-isDocument}{\code{geometa::ISOAbstractObject$isDocument()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isFieldInheritedFrom">}\href{../../geometa/html/ISOAbstractObject.html#method-isFieldInheritedFrom}{\code{geometa::ISOAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="print">}\href{../../geometa/html/ISOAbstractObject.html#method-print}{\code{geometa::ISOAbstractObject$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="save">}\href{../../geometa/html/ISOAbstractObject.html#method-save}{\code{geometa::ISOAbstractObject$save()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setAttr">}\href{../../geometa/html/ISOAbstractObject.html#method-setAttr}{\code{geometa::ISOAbstractObject$setAttr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeList">}\href{../../geometa/html/ISOAbstractObject.html#method-setCodeList}{\code{geometa::ISOAbstractObject$setCodeList()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeListValue">}\href{../../geometa/html/ISOAbstractObject.html#method-setCodeListValue}{\code{geometa::ISOAbstractObject$setCodeListValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeSpace">}\href{../../geometa/html/ISOAbstractObject.html#method-setCodeSpace}{\code{geometa::ISOAbstractObject$setCodeSpace()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setHref">}\href{../../geometa/html/ISOAbstractObject.html#method-setHref}{\code{geometa::ISOAbstractObject$setHref()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setId">}\href{../../geometa/html/ISOAbstractObject.html#method-setId}{\code{geometa::ISOAbstractObject$setId()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setIsNull">}\href{../../geometa/html/ISOAbstractObject.html#method-setIsNull}{\code{geometa::ISOAbstractObject$setIsNull()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setValue">}\href{../../geometa/html/ISOAbstractObject.html#method-setValue}{\code{geometa::ISOAbstractObject$setValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="validate">}\href{../../geometa/html/ISOAbstractObject.html#method-validate}{\code{geometa::ISOAbstractObject$validate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="wrapBaseElement">}\href{../../geometa/html/ISOAbstractObject.html#method-wrapBaseElement}{\code{geometa::ISOAbstractObject$wrapBaseElement()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decode"></a>}}
\if{latex}{\out{\hypertarget{method-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSGSElevationDomain$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-Class} from \pkg{XML}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSGSElevationDomain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
