% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSDescribeFeatureType.R
\docType{class}
\name{WFSDescribeFeatureType}
\alias{WFSDescribeFeatureType}
\title{WFSDescribeFeatureType}
\format{\code{\link{R6Class}} object.}
\usage{
WFSDescribeFeatureType
}
\value{
Object of \code{\link{R6Class}} for modelling a WFS DescribeFeatureType request
}
\description{
WFSDescribeFeatureType
}
\section{Methods}{

\describe{
 \item{\code{new(url, version, typeName)}}{
   This method is used to instantiate a WFSDescribeFeatureType object
 }
 \item{\code{getRequest()}}{
   Get DescribeFeatureType request
 }
 \item{\code{getContent()}}{
   Get content
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{DescribeFeatureType}
\keyword{OGC}
\keyword{WFS}
