% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSFeatureTypeElement.R
\docType{class}
\name{WFSFeatureTypeElement}
\alias{WFSFeatureTypeElement}
\title{WFSFeatureTypeElement}
\format{
\code{\link[R6]{R6Class}} object.
}
\value{
Object of \code{\link[R6]{R6Class}} modelling a WFS feature type element
}
\description{
WFSFeatureTypeElement

WFSFeatureTypeElement
}
\note{
Abstract class used by \pkg{ows4R}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{FeatureType}
\keyword{OGC}
\keyword{WFS}
\section{Super class}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{WFSFeatureTypeElement}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{minOccurs}}{minOccurs}

\item{\code{maxOccurs}}{maxOccurs}

\item{\code{nillable}}{nillable}

\item{\code{name}}{name}

\item{\code{type}}{type}

\item{\code{geometry}}{geometry}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WFSFeatureTypeElement-new}{\code{WFSFeatureTypeElement$new()}}
\item \href{#method-WFSFeatureTypeElement-getMinOccurs}{\code{WFSFeatureTypeElement$getMinOccurs()}}
\item \href{#method-WFSFeatureTypeElement-getMaxOccurs}{\code{WFSFeatureTypeElement$getMaxOccurs()}}
\item \href{#method-WFSFeatureTypeElement-isNillable}{\code{WFSFeatureTypeElement$isNillable()}}
\item \href{#method-WFSFeatureTypeElement-getName}{\code{WFSFeatureTypeElement$getName()}}
\item \href{#method-WFSFeatureTypeElement-getType}{\code{WFSFeatureTypeElement$getType()}}
\item \href{#method-WFSFeatureTypeElement-isGeometry}{\code{WFSFeatureTypeElement$isGeometry()}}
\item \href{#method-WFSFeatureTypeElement-clone}{\code{WFSFeatureTypeElement$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-ERROR'><code>ows4R::OGCAbstractObject$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-INFO'><code>ows4R::OGCAbstractObject$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-WARN'><code>ows4R::OGCAbstractObject$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-encode'><code>ows4R::OGCAbstractObject$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-getClass'><code>ows4R::OGCAbstractObject$getClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-getClassName'><code>ows4R::OGCAbstractObject$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-getNamespaceDefinition'><code>ows4R::OGCAbstractObject$getNamespaceDefinition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-isFieldInheritedFrom'><code>ows4R::OGCAbstractObject$isFieldInheritedFrom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-logger'><code>ows4R::OGCAbstractObject$logger()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="print"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-print'><code>ows4R::OGCAbstractObject$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WFSFeatureTypeElement-new"></a>}}
\if{latex}{\out{\hypertarget{method-WFSFeatureTypeElement-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{WFSFeatureTypeElement}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSFeatureTypeElement$new(xmlObj, namespaces)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmlObj}}{object of class \link[XML]{XMLInternalNode-class} from \pkg{XML}}

\item{\code{namespaces}}{namespaces definitions inherited from parent XML, as \code{data.frame}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WFSFeatureTypeElement-getMinOccurs"></a>}}
\if{latex}{\out{\hypertarget{method-WFSFeatureTypeElement-getMinOccurs}{}}}
\subsection{Method \code{getMinOccurs()}}{
get min occurs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSFeatureTypeElement$getMinOccurs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WFSFeatureTypeElement-getMaxOccurs"></a>}}
\if{latex}{\out{\hypertarget{method-WFSFeatureTypeElement-getMaxOccurs}{}}}
\subsection{Method \code{getMaxOccurs()}}{
get max occurs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSFeatureTypeElement$getMaxOccurs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WFSFeatureTypeElement-isNillable"></a>}}
\if{latex}{\out{\hypertarget{method-WFSFeatureTypeElement-isNillable}{}}}
\subsection{Method \code{isNillable()}}{
get if nillable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSFeatureTypeElement$isNillable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{logical}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WFSFeatureTypeElement-getName"></a>}}
\if{latex}{\out{\hypertarget{method-WFSFeatureTypeElement-getName}{}}}
\subsection{Method \code{getName()}}{
get name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSFeatureTypeElement$getName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WFSFeatureTypeElement-getType"></a>}}
\if{latex}{\out{\hypertarget{method-WFSFeatureTypeElement-getType}{}}}
\subsection{Method \code{getType()}}{
get type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSFeatureTypeElement$getType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WFSFeatureTypeElement-isGeometry"></a>}}
\if{latex}{\out{\hypertarget{method-WFSFeatureTypeElement-isGeometry}{}}}
\subsection{Method \code{isGeometry()}}{
Is geometry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSFeatureTypeElement$isGeometry()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{return}}{object of class \code{logical}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WFSFeatureTypeElement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WFSFeatureTypeElement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WFSFeatureTypeElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
