% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSGetFeature.R
\docType{class}
\name{WFSGetFeature}
\alias{WFSGetFeature}
\title{WFSGetFeature}
\format{\code{\link{R6Class}} object.}
\usage{
WFSGetFeature
}
\value{
Object of \code{\link{R6Class}} for modelling a WFS GetFeature request
}
\description{
WFSGetFeature
}
\note{
Abstract class used by \pkg{ows4R} to trigger a WFS DescribeFeatureType request
}
\section{Methods}{

\describe{
 \item{\code{new(op, url, version, typeName, ...)}}{
   This method is used to instantiate a WFSGetFeature object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GetFeature}
\keyword{OGC}
\keyword{WFS}
