% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WMSLayer.R
\docType{class}
\name{WMSLayer}
\alias{WMSLayer}
\title{WMSLayer}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WMS layer
}
\description{
WMSLayer
}
\note{
Abstract class used by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, capabilities, version, logger)}}{
   This method is used to instantiate a \code{WMSLayer} object
 }
 \item{\code{getName()}}{
   Get layer name
 }
 \item{\code{getTitle()}}{
   Get layer title
 }
 \item{\code{getAbstract()}}{
   Get layer abstract
 }
 \item{\code{getKeywords()}}{
   Get layer keywords
 }
 \item{\code{getDefaultCRS()}}{
   Get layer default CRS
 }
 \item{\code{getBoundingBox()}}{
   Get layer bounding box
 }
 \item{\code{getBoundingBoxSRS()}}{
   Get layer bounding box SRS
 }
 \item{\code{getBoundingBoxCRS()}}{
   Get layer bounding box CRS
 }
 \item{\code{getStyle()}}{
   Get layer style
 }
 \item{\code{getDimensions()}}{
   Get layer dimensions
 }
 \item{\code{getTimeDimension()}}{
   Get layer time dimension
 }
 \item{\code{getElevationDimension()}}{
   Get layer elevation dimension
 }
 \item{\code{getFeatureInfo(srs, styles, feature_count,
             x, y, width, height, bbox, info_format)}}{
   Get layer feature info                           
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Layer}
\keyword{OGC}
\keyword{WMS}
