% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSComplexOutputDescription.R
\docType{class}
\name{WPSComplexOutputDescription}
\alias{WPSComplexOutputDescription}
\title{WPSComplexOutputDescription}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process complex input description
}
\description{
WPSComplexOutputDescription
}
\note{
Class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xml, version, logger)}}{
   This method is used to instantiate a \code{WPSComplexOutputDescription} object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{WPS}
\keyword{complex}
\keyword{description}
\keyword{output}
