% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSDescribeProcess.R
\docType{class}
\name{WPSDescribeProcess}
\alias{WPSDescribeProcess}
\title{WPSDescribeProcess}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS DescribeProcess request
}
\description{
WPSDescribeProcess
}
\note{
Abstract class used by \pkg{ows4R} to trigger a WPS DescribeProcess request
}
\section{Methods}{

\describe{
 \item{\code{new(capabilities, op, url, version, identifier, logger, ...)}}{
   This method is used to instantiate a \code{WPSDescribeProcess} object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{DescribeProcess}
\keyword{OGC}
\keyword{WPS}
