% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSException.R
\docType{class}
\name{WPSException}
\alias{WPSException}
\title{WPSException}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS Exception
}
\description{
WPSException
}
\section{Methods}{

\describe{
 \item{\code{new(xml, serviceVersion)}}{
   This method is used to instantiate a \code{WPSException} object
 }
 \item{\code{decode(xml)}}{
   Decodes WPS exception from XML
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Exception}
\keyword{OGC}
\keyword{WPS}
