% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSExecute.R
\docType{class}
\name{WPSExecute}
\alias{WPSExecute}
\title{WPSExecute}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS Execute request
}
\description{
WPSExecute
}
\note{
Abstract class used by \pkg{ows4R} to trigger a WPS Execute request
}
\section{Methods}{

\describe{
 \item{\code{new(capabilities, op, url, serviceVersion, identifier, logger, ...)}}{
   This method is used to instantiate a WPSExecute object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Execute}
\keyword{OGC}
\keyword{WPS}
