% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSExecuteResponse.R
\docType{class}
\name{WPSExecuteResponse}
\alias{WPSExecuteResponse}
\title{WPSExecuteResponse}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS Execute response
}
\description{
WPSExecuteResponse

WPSExecuteResponse
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ExecuteResponse}
\keyword{OGC}
\keyword{WPS}
\section{Super class}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{WPSExecuteResponse}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{process}}{process}

\item{\code{status}}{status}

\item{\code{statusLocation}}{status location}

\item{\code{statusHistory}}{status history}

\item{\code{processOutputs}}{process outputs}

\item{\code{exception}}{exception}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WPSExecuteResponse$new()}}
\item \href{#method-getProcess}{\code{WPSExecuteResponse$getProcess()}}
\item \href{#method-getStatus}{\code{WPSExecuteResponse$getStatus()}}
\item \href{#method-getStatusLocation}{\code{WPSExecuteResponse$getStatusLocation()}}
\item \href{#method-getStatusHistory}{\code{WPSExecuteResponse$getStatusHistory()}}
\item \href{#method-getProcessOutputs}{\code{WPSExecuteResponse$getProcessOutputs()}}
\item \href{#method-getException}{\code{WPSExecuteResponse$getException()}}
\item \href{#method-decode}{\code{WPSExecuteResponse$decode()}}
\item \href{#method-update}{\code{WPSExecuteResponse$update()}}
\item \href{#method-clone}{\code{WPSExecuteResponse$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode">}\href{../../ows4R/html/OGCAbstractObject.html#method-encode}{\code{ows4R::OGCAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR">}\href{../../ows4R/html/OGCAbstractObject.html#method-ERROR}{\code{ows4R::OGCAbstractObject$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClass}{\code{ows4R::OGCAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClassName}{\code{ows4R::OGCAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition">}\href{../../ows4R/html/OGCAbstractObject.html#method-getNamespaceDefinition}{\code{ows4R::OGCAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO">}\href{../../ows4R/html/OGCAbstractObject.html#method-INFO}{\code{ows4R::OGCAbstractObject$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom">}\href{../../ows4R/html/OGCAbstractObject.html#method-isFieldInheritedFrom}{\code{ows4R::OGCAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger">}\href{../../ows4R/html/OGCAbstractObject.html#method-logger}{\code{ows4R::OGCAbstractObject$logger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN">}\href{../../ows4R/html/OGCAbstractObject.html#method-WARN}{\code{ows4R::OGCAbstractObject$WARN()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{WPSExecuteResponse}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$new(
  xml,
  capabilities,
  processDescription = NULL,
  logger = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{\code{capabilities}}{object of class \link{WPSCapabilities}}

\item{\code{processDescription}}{process description}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProcess"></a>}}
\if{latex}{\out{\hypertarget{method-getProcess}{}}}
\subsection{Method \code{getProcess()}}{
Get process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$getProcess()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{WPSProcess}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStatus"></a>}}
\if{latex}{\out{\hypertarget{method-getStatus}{}}}
\subsection{Method \code{getStatus()}}{
Get status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$getStatus()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{WPSStatus}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStatusLocation"></a>}}
\if{latex}{\out{\hypertarget{method-getStatusLocation}{}}}
\subsection{Method \code{getStatusLocation()}}{
Get status location
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$getStatusLocation()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStatusHistory"></a>}}
\if{latex}{\out{\hypertarget{method-getStatusHistory}{}}}
\subsection{Method \code{getStatusHistory()}}{
Get status history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$getStatusHistory()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProcessOutputs"></a>}}
\if{latex}{\out{\hypertarget{method-getProcessOutputs}{}}}
\subsection{Method \code{getProcessOutputs()}}{
Get list of process outputs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$getProcessOutputs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{list} of outputs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getException"></a>}}
\if{latex}{\out{\hypertarget{method-getException}{}}}
\subsection{Method \code{getException()}}{
Get exception
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$getException()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{WPSException}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decode"></a>}}
\if{latex}{\out{\hypertarget{method-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes an object of class \link{WPSExecuteResponse} from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$decode(xml, capabilities, processDescription, logger)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{\code{capabilities}}{object of class \link{WPSCapabilities}}

\item{\code{processDescription}}{process description}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Updates status history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$update(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{verbose. Default is \code{FALSE}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSExecuteResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
