% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSFormat.R
\docType{class}
\name{WPSFormat}
\alias{WPSFormat}
\title{WPSFormat}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process input description
}
\description{
WPSFormat
}
\note{
Class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, version, logger)}}{
   This method is used to instantiate a \code{WPSFormat} object
 }
 \item{\code{getMimeType()}}{
   Get mimetype
 }
 \item{\code{getEncoding()}}{
   Get encoding
 }
 \ìtem{\code{getSchema()}}{
   Get schema
 }
 \item{\code{setIsDefault(default)}}{
   Set if default format or not
 }
 \item{\code{isDefault()}}{
   Is default format
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{WPS}
\keyword{description}
\keyword{input}
