% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSLiteralOutputDescription.R
\docType{class}
\name{WPSLiteralOutputDescription}
\alias{WPSLiteralOutputDescription}
\title{WPSLiteralOutputDescription}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process literal output description
}
\description{
WPSLiteralOutputDescription
}
\note{
Class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xml, version, logger)}}{
   This method is used to instantiate a \code{WPSLiteralOutputDescription} object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{WPS}
\keyword{description}
\keyword{literal}
\keyword{output}
