% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSOutputDescription.R
\docType{class}
\name{WPSOutputDescription}
\alias{WPSOutputDescription}
\title{WPSOutputDescription}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process input description
}
\description{
WPSOutputDescription
}
\note{
Class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, version, logger)}}{
   This method is used to instantiate a \code{WPSOutputDescription} object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{WPS}
\keyword{description}
\keyword{output}
