% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSProcessDescription.R
\docType{class}
\name{WPSProcessDescription}
\alias{WPSProcessDescription}
\title{WPSProcessDescription}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process description
}
\description{
WPSProcessDescription
}
\note{
Class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, capabilities, version, logger)}}{
   This method is used to instantiate a \code{WPSProcessDescription} object
 }
 \item{\code{getIdentifier()}}{
   Get process identifier
 }
 \item{\code{getTitle()}}{
   Get process title
 }
 \item{\code{getAbstract()}}{
   Get process abstract
 }
 \item{\code{getVersion()}}{
   Get process version
 }
 \item{\code{isStatusSupported()}}{
   Get whether status is supported
 }
 \item{\code{isStoreSupported()}}{
   Get whether store is supported
 }
 \item{\code{asDataFrame()}}{
   Get process description as data.frame
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{ProcessDescription}
\keyword{WPS}
