% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSResponseDocument.R
\docType{class}
\name{WPSResponseDocument}
\alias{WPSResponseDocument}
\title{WPSResponseDocument}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS ResponseDocument
}
\description{
WPSResponseDocument
}
\section{Methods}{

\describe{
 \item{\code{new(xml, storeExecuteResponse, lineage, status, output)}}{
   This method is used to instantiate a WPSResponseDocument object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{ResponseDocument}
\keyword{WPS}
