% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSStatus.R
\docType{class}
\name{WPSStatus}
\alias{WPSStatus}
\title{WPSStatus}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS Status
}
\description{
WPSStatus
}
\section{Methods}{

\describe{
 \item{\code{new(xml, serviceVersion)}}{
   This method is used to instantiate a \code{WPSStatus} object
 }
 \item{\code{decode(xml)}}{
   Decodes WPS status from XML
 }
 \item{\code{getValue()}}{
   Get status value, among accepted WPS status values defined in the WPS standard:
   \code{ProcessAccepted}, \code{ProcessStarted}, \code{ProcessPaused}, \code{ProcessSucceeded},
   \code{ProcessFailed}.
 }
 \item{\code{getPercentCompleted()}}{
   Get percent completed
 }
 \item{\code{getCreationTime()}}{
   Get creation time
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Status}
\keyword{WPS}
