% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{And}
\alias{And}
\title{And}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an And operator
}
\description{
And
}
\section{Methods}{

\describe{
 \item{\code{new(...)}}{
   This method is used to instantiate an And operator
 }
}
}

\examples{
  expr1 <- PropertyIsEqualTo$new(PropertyName = "property1", Literal = "value1")
  expr2 <- PropertyIsEqualTo$new(PropertyName = "property2", Literal = "value2")
  and <- And$new(expr1,expr2)
  and_xml <- and$encode() #see how it looks like in XML

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{And}
\keyword{BinaryLogicOpType}
\keyword{Expression}
\keyword{OGC}
