% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{BBOX}
\alias{BBOX}
\title{BBOX}
\format{\code{\link{R6Class}} object.}
\usage{
BBOX
}
\value{
Object of \code{\link{R6Class}} for modelling an BBOX
}
\description{
BBOX
}
\section{Methods}{

\describe{
 \item{\code{new(bbox, srsName)}}{
   This method is used to instantiate an BBOX
 }
}
}

\keyword{BBOX}
\keyword{Expression}
\keyword{OGC}
