% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{BBox}
\alias{BBox}
\title{BBox}
\format{\code{\link{R6Class}} object.}
\usage{
BBox
}
\value{
Object of \code{\link{R6Class}} for modelling an BBox
}
\description{
BBox
}
\section{Methods}{

\describe{
 \item{\code{new(bbox, crs)}}{
   This method is used to instantiate an BBox
 }
}
}

\keyword{BBox}
\keyword{Expression}
\keyword{OGC}
