% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{BinaryComparisonOpType}
\alias{BinaryComparisonOpType}
\title{BinaryComparisonOpType}
\format{\code{\link{R6Class}} object.}
\usage{
BinaryComparisonOpType
}
\value{
Object of \code{\link{R6Class}} for modelling an BinaryComparisonOpType
}
\description{
BinaryComparisonOpType
}
\note{
abstract super class of all the property operation classes
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName, Literal, matchCase)}}{
   This method is used to instantiate an BinaryComparisonOpType
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{BinaryComparisonOpType}
\keyword{Expression}
\keyword{OGC}
