% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWClient.R
\docType{class}
\name{CSWClient}
\alias{CSWClient}
\title{CSWClient}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for interfacing an OGC
Catalogue Service for the Web.
}
\description{
CSWClient

CSWClient
}
\examples{
\donttest{
   #example based on CSW endpoint responding at http://localhost:8000/csw
   csw <- CSWClient$new("http://localhost:8000/csw", serviceVersion = "2.0.2")
   
   #get capabilities
   caps <- csw$getCapabilities()
   
   #get records
   records <- csw$getRecords()
   
   #get record by id
   record <- csw$getRecordById("my-metadata-id")
   
   #Advanced examples at https://github.com/eblondel/ows4R/wiki#csw
 }

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CSW}
\keyword{OGC}
\keyword{catalogue}
\keyword{service}
\keyword{web}
\section{Super classes}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{\link[ows4R:OWSClient]{ows4R::OWSClient}} -> \code{CSWClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CSWClient$new()}}
\item \href{#method-getCapabilities}{\code{CSWClient$getCapabilities()}}
\item \href{#method-reloadCapabilities}{\code{CSWClient$reloadCapabilities()}}
\item \href{#method-describeRecord}{\code{CSWClient$describeRecord()}}
\item \href{#method-getRecordById}{\code{CSWClient$getRecordById()}}
\item \href{#method-getRecords}{\code{CSWClient$getRecords()}}
\item \href{#method-transaction}{\code{CSWClient$transaction()}}
\item \href{#method-insertRecord}{\code{CSWClient$insertRecord()}}
\item \href{#method-updateRecord}{\code{CSWClient$updateRecord()}}
\item \href{#method-deleteRecord}{\code{CSWClient$deleteRecord()}}
\item \href{#method-deleteRecordById}{\code{CSWClient$deleteRecordById()}}
\item \href{#method-harvestRecord}{\code{CSWClient$harvestRecord()}}
\item \href{#method-harvestNode}{\code{CSWClient$harvestNode()}}
\item \href{#method-clone}{\code{CSWClient$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode">}\href{../../ows4R/html/OGCAbstractObject.html#method-encode}{\code{ows4R::OGCAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR">}\href{../../ows4R/html/OGCAbstractObject.html#method-ERROR}{\code{ows4R::OGCAbstractObject$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClass}{\code{ows4R::OGCAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClassName}{\code{ows4R::OGCAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition">}\href{../../ows4R/html/OGCAbstractObject.html#method-getNamespaceDefinition}{\code{ows4R::OGCAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO">}\href{../../ows4R/html/OGCAbstractObject.html#method-INFO}{\code{ows4R::OGCAbstractObject$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom">}\href{../../ows4R/html/OGCAbstractObject.html#method-isFieldInheritedFrom}{\code{ows4R::OGCAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger">}\href{../../ows4R/html/OGCAbstractObject.html#method-logger}{\code{ows4R::OGCAbstractObject$logger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN">}\href{../../ows4R/html/OGCAbstractObject.html#method-WARN}{\code{ows4R::OGCAbstractObject$WARN()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSClient" data-id="getCASUrl">}\href{../../ows4R/html/OWSClient.html#method-getCASUrl}{\code{ows4R::OWSClient$getCASUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSClient" data-id="getHeaders">}\href{../../ows4R/html/OWSClient.html#method-getHeaders}{\code{ows4R::OWSClient$getHeaders()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSClient" data-id="getPwd">}\href{../../ows4R/html/OWSClient.html#method-getPwd}{\code{ows4R::OWSClient$getPwd()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSClient" data-id="getToken">}\href{../../ows4R/html/OWSClient.html#method-getToken}{\code{ows4R::OWSClient$getToken()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSClient" data-id="getUrl">}\href{../../ows4R/html/OWSClient.html#method-getUrl}{\code{ows4R::OWSClient$getUrl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSClient" data-id="getUser">}\href{../../ows4R/html/OWSClient.html#method-getUser}{\code{ows4R::OWSClient$getUser()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSClient" data-id="getVersion">}\href{../../ows4R/html/OWSClient.html#method-getVersion}{\code{ows4R::OWSClient$getVersion()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
This method is used to instantiate a CSWClient with the \code{url} of the
   OGC service. Authentication is supported using basic auth (using \code{user}/\code{pwd} arguments), 
   bearer token (using \code{token} argument), or custom (using \code{headers} argument). By default, the \code{logger}
   argument will be set to \code{NULL} (no logger). This argument accepts two possible 
   values: \code{INFO}: to print only \pkg{ows4R} logs, \code{DEBUG}: to print more verbose logs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$new(
  url,
  serviceVersion = NULL,
  user = NULL,
  pwd = NULL,
  token = NULL,
  headers = c(),
  logger = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{url}

\item{\code{serviceVersion}}{CSW service version}

\item{\code{user}}{user}

\item{\code{pwd}}{password}

\item{\code{token}}{token}

\item{\code{headers}}{headers}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCapabilities"></a>}}
\if{latex}{\out{\hypertarget{method-getCapabilities}{}}}
\subsection{Method \code{getCapabilities()}}{
Get CSW capabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$getCapabilities()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{CSWCapabilities}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reloadCapabilities"></a>}}
\if{latex}{\out{\hypertarget{method-reloadCapabilities}{}}}
\subsection{Method \code{reloadCapabilities()}}{
Reloads CSW capabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$reloadCapabilities()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-describeRecord"></a>}}
\if{latex}{\out{\hypertarget{method-describeRecord}{}}}
\subsection{Method \code{describeRecord()}}{
Describe records. Retrieves the XML schema for CSW records. By default, returns the XML schema 
   for the CSW records (http://www.opengis.net/cat/csw/2.0.2). For other schemas, specify the
   \code{outputSchema} required, e.g. http://www.isotc211.org/2005/gmd for ISO 19115/19139 schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$describeRecord(namespace, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{namespace}

\item{\code{...}}{any other parameter to pass to the \link{CSWDescribeRecord} service request}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the service record description
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRecordById"></a>}}
\if{latex}{\out{\hypertarget{method-getRecordById}{}}}
\subsection{Method \code{getRecordById()}}{
Get a record by Id. By default, the record will be returned following the CSW schema 
   (http://www.opengis.net/cat/csw/2.0.2). For other schemas, specify the \code{outputSchema} 
   required,  e.g. http://www.isotc211.org/2005/gmd for ISO 19115/19139 records.
   The parameter \code{elementSetName} should among values "full", "brief", "summary". The default
   "full" corresponds to the full metadata sheet returned. "brief" and "summary" will contain only
   a subset of the metadata content.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$getRecordById(id, elementSetName = "full", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{record id}

\item{\code{elementSetName}}{element set name. Default is "full"}

\item{\code{...}}{any other parameter to pass to \link{CSWGetRecordById} service request}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the fetched record, \code{NULL} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRecords"></a>}}
\if{latex}{\out{\hypertarget{method-getRecords}{}}}
\subsection{Method \code{getRecords()}}{
Get records based on a query, object of class \code{CSWQuery}. The maximum number of records can be
   set either for the full query (\code{maxRecords}) or per request (\code{maxRecordsPerRequest}, default set to 10 records)
   considering this operation is paginated. By default, the record will be returned following the CSW schema 
   (http://www.opengis.net/cat/csw/2.0.2). For other schemas, specify the \code{outputSchema} 
   required,  e.g. http://www.isotc211.org/2005/gmd for ISO 19115/19139 records.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$getRecords(
  query = CSWQuery$new(),
  maxRecords = NULL,
  maxRecordsPerRequest = 10L,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{an object of class \link{CSWQuery}. By default, an empty query is set.}

\item{\code{maxRecords}}{max number of total records. Default is \code{NULL} meaning all records are returned.}

\item{\code{maxRecordsPerRequest}}{max number of records to return per request. Default set to 10.}

\item{\code{...}}{any other parameter to be passed to \link{CSWGetRecords} service request}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the list of records. By default each record will be returned as Dublin Core \code{list} object. In case 
 ISO 19115/19139 is set as \code{outputSchema}, each record will be object of class \code{ISOMetadata} from 
 \pkg{geometa}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transaction"></a>}}
\if{latex}{\out{\hypertarget{method-transaction}{}}}
\subsection{Method \code{transaction()}}{
Generic transaction method. Used for inserting, updating or deleting metadata using the transactional CSW service.
   The \code{type} gives the type of transaction (Insert, Update, or Delete). The record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$transaction(
  type,
  record = NULL,
  recordProperty = NULL,
  constraint = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{of transaction either "Insert", "Update" or "Delete"}

\item{\code{record}}{the record subject of the transaction}

\item{\code{recordProperty}}{record property, object of class \link{CSWRecordProperty}}

\item{\code{constraint}}{constraint, object of class \link{CSWConstraint}}

\item{\code{...}}{any other parameter to pass to \link{CSWTransaction} service request}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if transaction succeeded, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-insertRecord"></a>}}
\if{latex}{\out{\hypertarget{method-insertRecord}{}}}
\subsection{Method \code{insertRecord()}}{
Inserts a new record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$insertRecord(record, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{record}}{record subject of the Insertion}

\item{\code{...}}{any other parameter to pass to the transaction}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if insertion succeeded, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateRecord"></a>}}
\if{latex}{\out{\hypertarget{method-updateRecord}{}}}
\subsection{Method \code{updateRecord()}}{
Updates an existing record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$updateRecord(
  record = NULL,
  recordProperty = NULL,
  constraint = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{record}}{record subject of the Insertion}

\item{\code{recordProperty}}{record property, object of class \link{CSWRecordProperty}}

\item{\code{constraint}}{constraint, object of class \link{CSWConstraint}}

\item{\code{...}}{any other parameter to pass to the transaction}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if update succeeded, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deleteRecord"></a>}}
\if{latex}{\out{\hypertarget{method-deleteRecord}{}}}
\subsection{Method \code{deleteRecord()}}{
Deletes an existing (set of) record(s).  A constraint (object of class \code{CSWConstraint}) 
can be specified to limit the deletion to some records.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$deleteRecord(record = NULL, constraint = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{record}}{record subject of the Insertion}

\item{\code{constraint}}{constraint, object of class \link{CSWConstraint}}

\item{\code{...}}{any other parameter to pass to the transaction}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if deletion succeeded, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deleteRecordById"></a>}}
\if{latex}{\out{\hypertarget{method-deleteRecordById}{}}}
\subsection{Method \code{deleteRecordById()}}{
Deletes an existing record by identifier (constraint used to identify the record based on its identifier).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$deleteRecordById(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{record id}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if deletion succeeded, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-harvestRecord"></a>}}
\if{latex}{\out{\hypertarget{method-harvestRecord}{}}}
\subsection{Method \code{harvestRecord()}}{
Harvests a single record from a \code{sourceUrl}, given a \code{resourceType} (by default "http://www.isotc211.org/2005/gmd").
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$harvestRecord(
  sourceUrl,
  resourceType = "http://www.isotc211.org/2005/gmd"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sourceUrl}}{source URL}

\item{\code{resourceType}}{resource type. Default is "http://www.isotc211.org/2005/gmd"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if harvesting succeeded, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-harvestNode"></a>}}
\if{latex}{\out{\hypertarget{method-harvestNode}{}}}
\subsection{Method \code{harvestNode()}}{
Harvests a CSW node (having its endpoint defined by an  \code{url}). A \code{query} (object of class \code{CSWQuery}) can be
   specificed if needed to restrain the harvesting to a subset. The \code{resourceType} defines the type of resources to be harvested
   (by default "http://www.isotc211.org/2005/gmd")
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$harvestNode(
  url,
  query = CSWQuery$new(),
  resourceType = "http://www.isotc211.org/2005/gmd",
  sourceBaseUrl
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{CSW node URL}

\item{\code{query}}{a CSW query, object of class \link{CSWQuery}}

\item{\code{resourceType}}{resource type. Default is "http://www.isotc211.org/2005/gmd"}

\item{\code{sourceBaseUrl}}{source base URL}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class \code{list} giving the number of records \code{found} and those actually \code{harvested}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
