% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWConstraint.R
\docType{class}
\name{CSWConstraint}
\alias{CSWConstraint}
\title{CSWConstraint}
\format{\code{\link{R6Class}} object.}
\usage{
CSWConstraint
}
\value{
Object of \code{\link{R6Class}} for modelling an CSW Constraint
}
\description{
CSWConstraint
}
\section{Methods}{

\describe{
 \item{\code{new(cqlText, filter, serviceVersion)}}{
   This method is used to instantiate an CSWConstraint object.
 }
}
}

\examples{
  filter <- OGCFilter$new( PropertyIsEqualTo$new("apiso:Identifier", "12345") )
  cons <- CSWConstraint$new(filter = filter)
  cons_xml <- cons$encode() #how it looks like in XML

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Filter}
\keyword{OGC}
