% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWConstraint.R
\docType{class}
\name{CSWConstraint}
\alias{CSWConstraint}
\title{CSWConstraint}
\format{\code{\link{R6Class}} object.}
\usage{
CSWConstraint
}
\value{
Object of \code{\link{R6Class}} for modelling an CSW Constraint
}
\description{
CSWConstraint
}
\section{Methods}{

\describe{
 \item{\code{new(cqlText, filter, cswVersion)}}{
   This method is used to instantiate an CSWConstraint object.
 }
}
}

\keyword{Filter}
\keyword{OGC}
