% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWGetRecordById.R
\docType{class}
\name{CSWGetRecordById}
\alias{CSWGetRecordById}
\title{CSWGetRecordById}
\format{\code{\link{R6Class}} object.}
\usage{
CSWGetRecordById
}
\value{
Object of \code{\link{R6Class}} for modelling a CSW GetRecordById request
}
\description{
CSWGetRecordById
}
\note{
Class used internally by \pkg{ows4R} to trigger a CSW GetRecordById request
}
\section{Methods}{

\describe{
 \item{\code{new(op, url, serviceVersion, user, pwd, token, id, elementSetName, logger, ...)}}{
   This method is used to instantiate a CSWGetRecordById object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CSW}
\keyword{GetRecordById}
\keyword{OGC}
