% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWHarvest.R
\docType{class}
\name{CSWHarvest}
\alias{CSWHarvest}
\title{CSWHarvest}
\format{\code{\link{R6Class}} object.}
\usage{
CSWHarvest
}
\value{
Object of \code{\link{R6Class}} for modelling a CSW Harvest request
}
\description{
CSWHarvest
}
\section{Methods}{

\describe{
 \item{\code{new(op, url, serviceVersion, user, pwd, source, resourceType, resourceFormat, logger, ...)}}{
   This method is used to instantiate a CSWHarvest object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CSW}
\keyword{Harvest}
\keyword{OGC}
