% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWHarvest.R
\docType{class}
\name{CSWHarvest}
\alias{CSWHarvest}
\title{CSWHarvest}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a CSW Harvest request
}
\description{
CSWHarvest

CSWHarvest
}
\note{
Class used internally by \pkg{ows4R} to trigger a CSW Harvest request
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CSW}
\keyword{Harvest}
\keyword{OGC}
\section{Super classes}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{\link[ows4R:OWSHttpRequest]{ows4R::OWSHttpRequest}} -> \code{CSWHarvest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Source}}{source property for request XML encoding}

\item{\code{ResourceType}}{resource type property for request XML encoding}

\item{\code{ResourceFormat}}{resource format property for request XML encoding}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CSWHarvest$new()}}
\item \href{#method-clone}{\code{CSWHarvest$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode">}\href{../../ows4R/html/OGCAbstractObject.html#method-encode}{\code{ows4R::OGCAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR">}\href{../../ows4R/html/OGCAbstractObject.html#method-ERROR}{\code{ows4R::OGCAbstractObject$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClass}{\code{ows4R::OGCAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClassName}{\code{ows4R::OGCAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition">}\href{../../ows4R/html/OGCAbstractObject.html#method-getNamespaceDefinition}{\code{ows4R::OGCAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO">}\href{../../ows4R/html/OGCAbstractObject.html#method-INFO}{\code{ows4R::OGCAbstractObject$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom">}\href{../../ows4R/html/OGCAbstractObject.html#method-isFieldInheritedFrom}{\code{ows4R::OGCAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger">}\href{../../ows4R/html/OGCAbstractObject.html#method-logger}{\code{ows4R::OGCAbstractObject$logger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN">}\href{../../ows4R/html/OGCAbstractObject.html#method-WARN}{\code{ows4R::OGCAbstractObject$WARN()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSHttpRequest" data-id="execute">}\href{../../ows4R/html/OWSHttpRequest.html#method-execute}{\code{ows4R::OWSHttpRequest$execute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSHttpRequest" data-id="getCapabilities">}\href{../../ows4R/html/OWSHttpRequest.html#method-getCapabilities}{\code{ows4R::OWSHttpRequest$getCapabilities()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSHttpRequest" data-id="getException">}\href{../../ows4R/html/OWSHttpRequest.html#method-getException}{\code{ows4R::OWSHttpRequest$getException()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSHttpRequest" data-id="getRequest">}\href{../../ows4R/html/OWSHttpRequest.html#method-getRequest}{\code{ows4R::OWSHttpRequest$getRequest()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSHttpRequest" data-id="getRequestHeaders">}\href{../../ows4R/html/OWSHttpRequest.html#method-getRequestHeaders}{\code{ows4R::OWSHttpRequest$getRequestHeaders()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSHttpRequest" data-id="getResponse">}\href{../../ows4R/html/OWSHttpRequest.html#method-getResponse}{\code{ows4R::OWSHttpRequest$getResponse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSHttpRequest" data-id="getResult">}\href{../../ows4R/html/OWSHttpRequest.html#method-getResult}{\code{ows4R::OWSHttpRequest$getResult()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSHttpRequest" data-id="getStatus">}\href{../../ows4R/html/OWSHttpRequest.html#method-getStatus}{\code{ows4R::OWSHttpRequest$getStatus()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OWSHttpRequest" data-id="setResult">}\href{../../ows4R/html/OWSHttpRequest.html#method-setResult}{\code{ows4R::OWSHttpRequest$setResult()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{CSWHarvest} service request
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWHarvest$new(
  capabilities,
  op,
  url,
  serviceVersion = "2.0.2",
  user = NULL,
  pwd = NULL,
  token = NULL,
  headers = list(),
  source = NULL,
  resourceType = "http://www.isotc211.org/schemas/2005/gmd/",
  resourceFormat = "application/xml",
  logger = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{capabilities}}{an object of class \link{CSWCapabilities}}

\item{\code{op}}{object of class \link{OWSOperation} as retrieved from capabilities}

\item{\code{url}}{url}

\item{\code{serviceVersion}}{serviceVersion. Default is "2.0.2}

\item{\code{user}}{user}

\item{\code{pwd}}{password}

\item{\code{token}}{token}

\item{\code{headers}}{headers}

\item{\code{source}}{source}

\item{\code{resourceType}}{resource type. Default is "http://www.isotc211.org/schemas/2005/gmd/"}

\item{\code{resourceFormat}}{resource format. Default is "application/xml"}

\item{\code{logger}}{logger}

\item{\code{...}}{any parameter to pass to the service request}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSWHarvest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
