% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWQuery.R
\docType{class}
\name{CSWQuery}
\alias{CSWQuery}
\title{CSWQuery}
\format{\code{\link{R6Class}} object.}
\usage{
CSWQuery
}
\value{
Object of \code{\link{R6Class}} for modelling an CSW Query
}
\description{
CSWQuery
}
\section{Methods}{

\describe{
 \item{\code{new(filter, serviceVersion)}}{
   This method is used to instantiate an CSWQUery object.
 }
}
}

\keyword{OGC}
\keyword{Query}
