% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWRecordProperty.R
\docType{class}
\name{CSWRecordProperty}
\alias{CSWRecordProperty}
\title{CSWRecordProperty}
\format{\code{\link{R6Class}} object.}
\usage{
CSWRecordProperty
}
\value{
Object of \code{\link{R6Class}} for modelling an CSW RecordProperty
}
\description{
CSWRecordProperty
}
\section{Methods}{

\describe{
 \item{\code{new(name, value)}}{
   This method is used to instantiate an CSWRecordProperty object.
 }
}
}

\keyword{CSW}
\keyword{RecordProperty}
