% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCAbstractObject.R
\docType{class}
\name{OGCAbstractObject}
\alias{OGCAbstractObject}
\title{OGCAbstractObject}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OGCAbstractObject
}
\description{
OGCAbstractObject
}
\note{
abstract class used by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, namespacePrefix, attrs, defaults, wrap, logger)}}{
   This method is used to instantiate an OGCAbstractObject
 }
 \item{\code{getClassName()}}{
   Get class name
 }
 \item{\code{getClass()}}{
   Get class
 }
 \item{\code{encode(addNS, geometa_validate, geometa_inspire, geometa_inspireValidator)}}{
   Encodes as XML. The \code{addNS} controls the addition of XML namespaces.
   Extra parameters related to \pkg{geometa} objects: \code{geometa_validate} (TRUE by default) and \code{geometa_inspire} 
   (FALSE by default) can be used to perform ISO and INSPIRE validation respectively. In that case on object of class 
   \code{geometa::INSPIREMetadataValidator}, with a proper user API key, should be specified as \code{geometa_inspireValidator} 
   argument.
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Abstract}
\keyword{OGC}
\keyword{Object}
