% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{OGCExpression}
\alias{OGCExpression}
\title{OGCExpression}
\format{\code{\link{R6Class}} object.}
\usage{
OGCExpression
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Expression
}
\description{
OGCExpression
}
\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new()}}{
   This method is used to instantiate an OGCExpression object
 }
}
}

\keyword{Expression}
\keyword{OGC}
