% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCFilter.R
\docType{class}
\name{OGCFilter}
\alias{OGCFilter}
\title{OGCFilter}
\format{\code{\link{R6Class}} object.}
\usage{
OGCFilter
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Filter
}
\description{
OGCFilter
}
\section{Methods}{

\describe{
 \item{\code{new(expr)}}{
   This method is used to instantiate an OGCFilter object. The unique
   argument should be an object of class \code{\link{OGCExpression}}
 }
}
}

\keyword{Filter}
\keyword{OGC}
