% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCFilter.R
\docType{class}
\name{OGCFilter}
\alias{OGCFilter}
\title{OGCFilter}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Filter
}
\description{
OGCFilter
}
\section{Methods}{

\describe{
 \item{\code{new(exprn filterVersion)}}{
   This method is used to instantiate an OGCFilter object. The unique
   argument should be an object of class \code{\link{OGCExpression}}
 }
 \item{\code{setFilterVersion(filterVersion)}}{
   Set filter version
 }
}
}

\examples{
  expr <- PropertyIsEqualTo$new(PropertyName = "property", Literal = "value")
  not <- Not$new(expr)
  not_xml <- not$encode() #see how it looks like in XML

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Filter}
\keyword{OGC}
