% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{Or}
\alias{Or}
\title{Or}
\format{\code{\link{R6Class}} object.}
\usage{
Or
}
\value{
Object of \code{\link{R6Class}} for modelling an Or operator
}
\description{
Or
}
\section{Methods}{

\describe{
 \item{\code{new(...)}}{
   This method is used to instantiate an Or operator
 }
}
}

\examples{
  expr1 <- PropertyIsEqualTo$new(PropertyName = "property1", Literal = "value1")
  expr2 <- PropertyIsEqualTo$new(PropertyName = "property2", Literal = "value2")
  or <- Or$new(expr1,expr2)
  or_xml <- or$encode() #see how it looks like in XML

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{BinaryLogicOpType}
\keyword{Expression}
\keyword{OGC}
\keyword{Or}
