% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4R.R
\docType{package}
\name{ows4R}
\alias{ows4R}
\alias{ows4R-package}
\alias{ows4R-package}
\title{Interface to OGC Web-Services}
\description{
Provides an interface to OGC Web-Services (OWS). In a first step, the package supports the Common
OGC Web-Services specifications the Web Feature Service (WFS). ows4R will progressively support other OGC web
service specifications such as Web Map Service (WMS), Web Coverage Service, or Catalogue Service for the web (CSW).
}
\details{
\tabular{ll}{
   Package: \tab ows4R\cr
   Type: \tab Package\cr
   Version
   : \tab 0.1-1\cr
   Date: \tab 2018-05-18\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
