% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ows4R.R
\docType{package}
\name{ows4R}
\alias{ows4R}
\alias{ows4R-package}
\title{Interface to OGC Web-Services}
\description{
Provides an Interface to Web-Services defined as standards by the Open Geospatial Consortium (OGC), including Web Feature Service
(WFS) for vector data, Catalogue Service (CSW) for ISO/OGC metadata, and associated standards such as the common web-service specification (OWS) and
OGC Filter Encoding. Partial support is provided for the Web Map Service (WMS) and Web Processing Service (WPS). The purpose is to add support for 
additional OGC service standards such as Web Coverage Service (WCS), Web Coverage Processing Service (WCPS) or OGC API.
}
\details{
\tabular{ll}{
   Package: \tab ows4R\cr
   Type: \tab Package\cr
   Version
   : \tab 0.2\cr
   Date: \tab 2021-07-090\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
