% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSClient.R
\docType{class}
\name{OWSClient}
\alias{OWSClient}
\title{OWSClient}
\format{\code{\link{R6Class}} object.}
\usage{
OWSClient
}
\value{
Object of \code{\link{R6Class}} with methods for interfacing
a Common OGC web-service.
}
\description{
OWSClient
}
\section{Fields}{

\describe{
\item{\code{loggerType}}{the type of logger}

\item{\code{verbose.info}}{if basic logs have to be printed}

\item{\code{verbose.debug}}{if verbose logs have to be printed}

\item{\code{url}}{the Base url of OWS service}

\item{\code{version}}{the version of OWS service}
}}

\section{Methods}{

\describe{
 \item{\code{new(url, version, user, pwd, logger)}}{
   This method is used to instantiate a OWSClient with the \code{url} of the
   OGC service. Authentication (\code{user}/\code{pwd}) is not yet supported and will
   be added with the support of service transactional modes. By default, the \code{logger}
   argument will be set to \code{NULL} (no logger). This argument accepts two possible 
   values: \code{INFO}: to print only \pkg{ows4R} logs, \code{DEBUG}: to print more verbose logs
 }
 \item{\code{logger(type, text)}}{
   Basic logger to report logs. Used internally
 }
 \item{\code{INFO(text)}}{
   Logger to report information. Used internally
 }
 \item{\code{WARN(text)}}{
   Logger to report warnings. Used internally
 }
 \item{\code{ERROR(text)}}{
   Logger to report errors. Used internally
 }
 \item{\code{getUrl()}}{
   Get the service URL
 }
 \item{\code{getVersion()}}{
   Get the service version
 }
 \item{\code{getCapabilities()}}{
   Get the service capabilities
 }
 \item{\code{getClassName()}}{
   Retrieves the name of the class instance
 }
}
}

\examples{
\dontrun{
   OWSClient$new("http://localhost:8080/geoserver/ows", version = "1.1.0")
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Common}
\keyword{OGC}
\keyword{OWS}
