% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSGetCapabilities.R
\docType{class}
\name{OWSGetCapabilities}
\alias{OWSGetCapabilities}
\title{OWSGetCapabilities}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GetCapabilities request
}
\description{
OWSGetCapabilities
}
\note{
Abstract class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(capabilities, op, url, service, version, ...)}}{
   This method is used to instantiate a OWSGetCapabilities object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GetCapabilities}
\keyword{OGC}
