% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSNamespace.R
\docType{class}
\name{OWSNamespace}
\alias{OWSNamespace}
\title{OWSNamespace}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OWS Namespace
}
\description{
OWSNamespace

OWSNamespace
}
\note{
class used internally by ows4R for specifying XML namespaces
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OWS}
\keyword{namespace}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{namespace id}

\item{\code{uri}}{namespace uri}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OWSNamespace$new()}}
\item \href{#method-getDefinition}{\code{OWSNamespace$getDefinition()}}
\item \href{#method-clone}{\code{OWSNamespace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes an \link{OWSNamespace}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSNamespace$new(id, uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{uri}}{uri}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDefinition"></a>}}
\if{latex}{\out{\hypertarget{method-getDefinition}{}}}
\subsection{Method \code{getDefinition()}}{
Get namespace definition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSNamespace$getDefinition()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a named \code{list} with id and uri
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OWSNamespace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
