% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSOperation.R
\docType{class}
\name{OWSOperation}
\alias{OWSOperation}
\title{OWSOperation}
\format{\code{\link{R6Class}} object.}
\usage{
OWSOperation
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Operation
}
\description{
OWSOperation
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, service, version)}}{
   This method is used to instantiate an OWSOperation object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{OWS}
\keyword{operation}
