% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSOperation.R
\docType{class}
\name{OWSOperation}
\alias{OWSOperation}
\title{OWSOperation}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Operation
}
\description{
OWSOperation
}
\note{
Abstract class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, serviceVersion)}}{
   This method is used to instantiate an OWSOperation object
 }
 \item{\code{getName()}}{
   Get name
 }
 \item{\code{getParameters()}}{
   Get the list of parameters
 }
 \item{\code{getParameter(name)}}{
   Get a given parameter
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{OWS}
\keyword{operation}
