% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSOperationsMetadata.R
\docType{class}
\name{OWSOperationsMetadata}
\alias{OWSOperationsMetadata}
\title{OWSOperationsMetadata}
\format{\code{\link{R6Class}} object.}
\usage{
OWSOperationsMetadata
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Operations Metadata
}
\description{
OWSOperationsMetadata
}
\note{
Abstract class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, serviceVersion)}}{
   This method is used to instantiate a OWSOperationsMetadata object
 }
 \item{\code{getOperations()}}{
   Get operations
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{OWS}
\keyword{metadata}
\keyword{operation}
