% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSRequest.R
\docType{class}
\name{OWSRequest}
\alias{OWSRequest}
\title{OWSRequest}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a OWS Service Capability Request
}
\description{
OWSRequest
}
\note{
Abstract class used by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, capabilities, version, logger)}}{
   This method is used to instantiate a \code{OWSRequest} object
 }
 \item{\code{getName()}}{
   Get request name
 }
 \item{\code{getFormats()}}{
   Get request formats
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Request}
\keyword{Service}
