% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSRequest.R
\docType{class}
\name{OWSRequest}
\alias{OWSRequest}
\title{OWSRequest}
\format{\code{\link{R6Class}} object.}
\usage{
OWSRequest
}
\value{
Object of \code{\link{R6Class}} for modelling a generic OWS request
}
\description{
OWSRequest
}
\section{Fields}{

\describe{
\item{\code{response}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(op, type, url, request, user, pwd, namedParams, attrs, 
                 contentType, mimeType, logger)}}{
   This method is used to instantiate a object for doing an OWS request
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{OWS}
\keyword{Request}
