% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSServiceProvider.R
\docType{class}
\name{OWSServiceProvider}
\alias{OWSServiceProvider}
\title{OWSServiceProvider}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC Service Provider
}
\description{
OWSServiceProvider
}
\note{
Abstract class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, owsVersion, serviceVersion)}}{
   This method is used to instantiate a OWSServiceProvider object
 }
 \item{\code{getProviderName()}}{
   Get the provider name
 }
 \item{\code{getProviderSite()}}{
   Get the provide site
 }
 \item{\code{getServiceContact()}}{
   Get the service contact, as object of class \code{ISOResponsibleParty}
   from package \pkg{geometa}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{OWS}
\keyword{Provider}
\keyword{Service}
