% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWSWGS84BoundingBox.R
\docType{class}
\name{OWSWGS84BoundingBox}
\alias{OWSWGS84BoundingBox}
\title{OWSWGS84BoundingBox}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC WS84 BoundingBox
}
\description{
OWSWGS84BoundingBox
}
\note{
Class used internally by \pkg{geometa}
}
\section{Methods}{

\describe{
 \item{\code{new(xml, owsVersion, serviceVersion, logger)}}{
   This method is used to instantiate an \code{OWSWGS84BoundingBox} object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{OWS}
\keyword{boundingbox}
\keyword{wgs84}
