% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsGreaterThanOrEqualTo}
\alias{PropertyIsGreaterThanOrEqualTo}
\title{PropertyIsGreaterThanOrEqualTo}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsGreaterThanOrEqualTo
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsGreaterThanOrEqualTo
}
\description{
PropertyIsGreaterThanOrEqualTo
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName, Literal, matchCase)}}{
   This method is used to instantiate an PropertyIsGreaterThanOrEqualTo
 }
}
}

\keyword{BinaryComparisonOpType}
\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsGreaterThanOrEqualTo}
