% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsLessThan}
\alias{PropertyIsLessThan}
\title{PropertyIsLessThan}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsLessThan
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsLessThan
}
\description{
PropertyIsLessThan
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName, Literal, matchCase)}}{
   This method is used to instantiate an PropertyIsLessThan
 }
}
}

\keyword{BinaryComparisonOpType}
\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsLessThan}
