% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsLessThanOrEqualTo}
\alias{PropertyIsLessThanOrEqualTo}
\title{PropertyIsLesserThanOrEqualTo}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsLessThanOrEqualTo
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsLesserThanOrEqualTo
}
\description{
PropertyIsLesserThanOrEqualTo
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName, Literal, matchCase)}}{
   This method is used to instantiate an PropertyIsLesserThanOrEqualTo
 }
}
}

\keyword{BinaryComparisonOpType}
\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsLesserThanOrEqualTo}
