% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsLike}
\alias{PropertyIsLike}
\title{PropertyIsLike}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsLike
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsLike
}
\description{
PropertyIsLike
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyOperator, PropertyName, Literal,
                 escapeChar, singleChar, wildCard, matchCase)}}{
   This method is used to instantiate an PropertyIsLike
 }
}
}

\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsLike}
