% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsNotEqualTo}
\alias{PropertyIsNotEqualTo}
\title{PropertyIsNotEqualTo}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsNotEqualTo
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsNotEqualTo
}
\description{
PropertyIsNotEqualTo
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName, Literal, matchCase)}}{
   This method is used to instantiate an PropertyIsNotEqualTo
 }
}
}

\examples{
  expr <- PropertyIsNotEqualTo$new(PropertyName = "property", Literal = "value")
  expr_xml <- expr$encode() #see how it looks like in XML

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{BinaryComparisonOpType}
\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsNotEqualTo}
