% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{PropertyIsNull}
\alias{PropertyIsNull}
\title{PropertyIsNull}
\format{\code{\link{R6Class}} object.}
\usage{
PropertyIsNull
}
\value{
Object of \code{\link{R6Class}} for modelling an PropertyIsNull
}
\description{
PropertyIsNull
}
\section{Methods}{

\describe{
 \item{\code{new(PropertyName)}}{
   This method is used to instantiate an PropertyIsNull
 }
}
}

\keyword{Expression}
\keyword{OGC}
\keyword{PropertyIsNull}
