% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGCExpression.R
\docType{class}
\name{UnaryLogicOpType}
\alias{UnaryLogicOpType}
\title{UnaryLogicOpType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an UnaryLogicOpType
}
\description{
UnaryLogicOpType
}
\note{
abstract super class of all the unary logical operation classes
}
\section{Methods}{

\describe{
 \item{\code{new(...)}}{
   This method is used to instantiate an UnaryLogicOpType
 }
}
}

\keyword{Expression}
\keyword{OGC}
\keyword{UnaryLogicOpType}
