% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCSCoverageDomain.R
\docType{class}
\name{WCSCoverageDomain}
\alias{WCSCoverageDomain}
\title{WCSCoverageDomain}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WCS coverage domain
}
\description{
WCSCoverageDomain

WCSCoverageDomain
}
\note{
Class used internally by ows4R.
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Coverage}
\keyword{OGC}
\keyword{WCS}
\keyword{domain}
\section{Super class}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{WCSCoverageDomain}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{spatialDomain}}{spatial domain}

\item{\code{temporalDomain}}{temporal domain}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WCSCoverageDomain-new}{\code{WCSCoverageDomain$new()}}
\item \href{#method-WCSCoverageDomain-getSpatialDomain}{\code{WCSCoverageDomain$getSpatialDomain()}}
\item \href{#method-WCSCoverageDomain-getTemporalDomain}{\code{WCSCoverageDomain$getTemporalDomain()}}
\item \href{#method-WCSCoverageDomain-clone}{\code{WCSCoverageDomain$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-ERROR'><code>ows4R::OGCAbstractObject$ERROR()</code></a></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-INFO'><code>ows4R::OGCAbstractObject$INFO()</code></a></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-WARN'><code>ows4R::OGCAbstractObject$WARN()</code></a></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-encode'><code>ows4R::OGCAbstractObject$encode()</code></a></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-getClass'><code>ows4R::OGCAbstractObject$getClass()</code></a></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-getClassName'><code>ows4R::OGCAbstractObject$getClassName()</code></a></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-getNamespaceDefinition'><code>ows4R::OGCAbstractObject$getNamespaceDefinition()</code></a></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-isFieldInheritedFrom'><code>ows4R::OGCAbstractObject$isFieldInheritedFrom()</code></a></li>
<li><span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger"><a href='../../ows4R/html/OGCAbstractObject.html#method-OGCAbstractObject-logger'><code>ows4R::OGCAbstractObject$logger()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverageDomain-new"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverageDomain-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{WCSCoverageDomain}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDomain$new(xmlObj, serviceVersion, owsVersion, logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmlObj}}{an object of class \link{XMLInternalNode-class} to initialize from XML}

\item{\code{serviceVersion}}{service version}

\item{\code{owsVersion}}{OWS version}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverageDomain-getSpatialDomain"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverageDomain-getSpatialDomain}{}}}
\subsection{Method \code{getSpatialDomain()}}{
Get spatial domain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDomain$getSpatialDomain()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \link{WCSCoverageSpatialDomain}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverageDomain-getTemporalDomain"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverageDomain-getTemporalDomain}{}}}
\subsection{Method \code{getTemporalDomain()}}{
Get spatial domain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDomain$getTemporalDomain()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \link{WCSCoverageTemporalDomain}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WCSCoverageDomain-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WCSCoverageDomain-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageDomain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
