% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCSCoverageSummary.R
\docType{class}
\name{WCSCoverageSummary}
\alias{WCSCoverageSummary}
\title{WCSCoverageSummary}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WCS coverage summary
}
\description{
WCSCoverageSummary

WCSCoverageSummary
}
\note{
Class used internally by ows4R.
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Coverage}
\keyword{OGC}
\keyword{WCS}
\section{Super class}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{WCSCoverageSummary}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CoverageId}}{coverage id}

\item{\code{CoverageSubtype}}{coverage subtype}

\item{\code{CoverageSubtypeParent}}{coverage subtype parent}

\item{\code{WGS84BoundingBox}}{WGS84 bounding box}

\item{\code{BoundingBox}}{bounding box}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WCSCoverageSummary$new()}}
\item \href{#method-getId}{\code{WCSCoverageSummary$getId()}}
\item \href{#method-getSubtype}{\code{WCSCoverageSummary$getSubtype()}}
\item \href{#method-getSubtypeParent}{\code{WCSCoverageSummary$getSubtypeParent()}}
\item \href{#method-getWGS84BoundingBox}{\code{WCSCoverageSummary$getWGS84BoundingBox()}}
\item \href{#method-getBoundingBox}{\code{WCSCoverageSummary$getBoundingBox()}}
\item \href{#method-getDescription}{\code{WCSCoverageSummary$getDescription()}}
\item \href{#method-getDimensions}{\code{WCSCoverageSummary$getDimensions()}}
\item \href{#method-getCoverage}{\code{WCSCoverageSummary$getCoverage()}}
\item \href{#method-getCoverageStack}{\code{WCSCoverageSummary$getCoverageStack()}}
\item \href{#method-clone}{\code{WCSCoverageSummary$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode">}\href{../../ows4R/html/OGCAbstractObject.html#method-encode}{\code{ows4R::OGCAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR">}\href{../../ows4R/html/OGCAbstractObject.html#method-ERROR}{\code{ows4R::OGCAbstractObject$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClass}{\code{ows4R::OGCAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClassName}{\code{ows4R::OGCAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition">}\href{../../ows4R/html/OGCAbstractObject.html#method-getNamespaceDefinition}{\code{ows4R::OGCAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO">}\href{../../ows4R/html/OGCAbstractObject.html#method-INFO}{\code{ows4R::OGCAbstractObject$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom">}\href{../../ows4R/html/OGCAbstractObject.html#method-isFieldInheritedFrom}{\code{ows4R::OGCAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger">}\href{../../ows4R/html/OGCAbstractObject.html#method-logger}{\code{ows4R::OGCAbstractObject$logger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN">}\href{../../ows4R/html/OGCAbstractObject.html#method-WARN}{\code{ows4R::OGCAbstractObject$WARN()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{WCSCoverageSummary} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$new(
  xmlObj,
  capabilities,
  serviceVersion,
  owsVersion,
  logger = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xmlObj}}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{\code{capabilities}}{object of class \link{WCSCapabilities}}

\item{\code{serviceVersion}}{WCS service version}

\item{\code{owsVersion}}{version}

\item{\code{logger}}{logger type \code{NULL}, "INFO" or "DEBUG"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getId"></a>}}
\if{latex}{\out{\hypertarget{method-getId}{}}}
\subsection{Method \code{getId()}}{
Get coverage ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getId()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSubtype"></a>}}
\if{latex}{\out{\hypertarget{method-getSubtype}{}}}
\subsection{Method \code{getSubtype()}}{
Get sub type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getSubtype()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSubtypeParent"></a>}}
\if{latex}{\out{\hypertarget{method-getSubtypeParent}{}}}
\subsection{Method \code{getSubtypeParent()}}{
Get sub type parent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getSubtypeParent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getWGS84BoundingBox"></a>}}
\if{latex}{\out{\hypertarget{method-getWGS84BoundingBox}{}}}
\subsection{Method \code{getWGS84BoundingBox()}}{
Get bounding box
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getWGS84BoundingBox()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{OWSWGS84BoundingBox}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBoundingBox"></a>}}
\if{latex}{\out{\hypertarget{method-getBoundingBox}{}}}
\subsection{Method \code{getBoundingBox()}}{
Get WGS84 bounding box
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getBoundingBox()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{OWSBoundingBox}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDescription"></a>}}
\if{latex}{\out{\hypertarget{method-getDescription}{}}}
\subsection{Method \code{getDescription()}}{
Get description
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getDescription()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link{WCSCoverageDescription}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDimensions"></a>}}
\if{latex}{\out{\hypertarget{method-getDimensions}{}}}
\subsection{Method \code{getDimensions()}}{
Get dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getDimensions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the list of dimensions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCoverage"></a>}}
\if{latex}{\out{\hypertarget{method-getCoverage}{}}}
\subsection{Method \code{getCoverage()}}{
Get coverage data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getCoverage(
  bbox = NULL,
  crs = NULL,
  time = NULL,
  elevation = NULL,
  format = NULL,
  rangesubset = NULL,
  gridbaseCRS = NULL,
  gridtype = NULL,
  gridCS = NULL,
  gridorigin = NULL,
  gridoffsets = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bbox}}{bbox. Default is \code{NULL}}

\item{\code{crs}}{crs. Default is \code{NULL}}

\item{\code{time}}{time. Default is \code{NULL}}

\item{\code{elevation}}{elevation. Default is \code{NULL}}

\item{\code{format}}{format. Default will be GeoTIFF, coded differently depending on the WCS version.}

\item{\code{rangesubset}}{rangesubset. Default is \code{NULL}}

\item{\code{gridbaseCRS}}{grid base CRS. Default is \code{NULL}}

\item{\code{gridtype}}{grid type. Default is \code{NULL}}

\item{\code{gridCS}}{grid CS. Default is \code{NULL}}

\item{\code{gridorigin}}{grid origin. Default is \code{NULL}}

\item{\code{gridoffsets}}{grid offsets. Default is \code{NULL}}

\item{\code{...}}{any other argument to \link{WCSGetCoverage}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class \link{raster} from \pkg{raster}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCoverageStack"></a>}}
\if{latex}{\out{\hypertarget{method-getCoverageStack}{}}}
\subsection{Method \code{getCoverageStack()}}{
Get a spatio-temporal coverage data cubes as coverage \link{stack}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$getCoverageStack(time = NULL, elevation = NULL, bbox = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{time}

\item{\code{elevation}}{elevation}

\item{\code{bbox}}{bbox}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an object of class \link{stack} from \pkg{raster}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WCSCoverageSummary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
