% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSCapabilities.R
\docType{class}
\name{WFSCapabilities}
\alias{WFSCapabilities}
\title{WFSGetCapabilities}
\format{\code{\link{R6Class}} object.}
\usage{
WFSCapabilities
}
\value{
Object of \code{\link{R6Class}} with methods for interfacing an OGC
Web Feature Service Get Capabilities document.
}
\description{
WFSGetCapabilities
}
\section{Methods}{

\describe{
 \item{\code{new(url, version)}}{
   This method is used to instantiate a WFSGetCapabilities object
 }
 \item{\code{getServiceIdentification()}}{
   Get the service identification
 }
 \item{\code{getFeatureTypes()}}{
   Retrieves the list of feature types
 }
 \item{\code{findFeatureTypeByName(name, exact)}}{
   Find feature type(s) by name.
 }
}
}

\examples{
\dontrun{
   WFSGetCapabilities$new("http://localhost:8080/geoserver/wfs", version = "1.1.1")
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GetCapabilities}
\keyword{OGC}
\keyword{WFS}
