% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFSFeatureTypeElement.R
\docType{class}
\name{WFSFeatureTypeElement}
\alias{WFSFeatureTypeElement}
\title{WFSFeatureTypeElement}
\format{\code{\link{R6Class}} object.}
\usage{
WFSFeatureTypeElement
}
\value{
Object of \code{\link{R6Class}} modelling a WFS feature type element
}
\description{
WFSFeatureTypeElement
}
\note{
Class used internally by ows4R.
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj)}}{
   This method is used to instantiate a \code{WFSFeatureTypeElement} object
 }
 \item{\code{getMinOccurs()}}{
   Get min occurs
 }
 \item{\code{getMaxOccurs()}}{
   Get max occurs
 }
 \item{\code{isNillable()}}{
   Get TRUE if nillable, FALSE otherwise
 }
 \item{\code{getName()}}{
   Get element name
 }
 \item{\code{getType()}}{
   Get element type
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{FeatureType}
\keyword{OGC}
\keyword{WFS}
