% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WMSGetFeatureInfo.R
\docType{class}
\name{WMSGetFeatureInfo}
\alias{WMSGetFeatureInfo}
\title{WMSGetFeatureInfo}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WMS GetFeatureInfo request
}
\description{
WMSGetFeatureInfo
}
\note{
Abstract class used by \pkg{ows4R} to trigger a WMS GetFeatureInfo request
}
\section{Methods}{

\describe{
 \item{\code{new(op, url, version, layers, srs, styles, feature_count,
                 x, y, width, height, bbox, info_format, logger, ...)}}{
   This method is used to instantiate a WMSGetFeatureInfo object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GetFeatureInfo}
\keyword{OGC}
\keyword{WMS}
