% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WMSLayer.R
\docType{class}
\name{WMSLayer}
\alias{WMSLayer}
\title{WMSLayer}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WMS layer
}
\description{
WMSLayer
}
\note{
Abstract class used by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, capabilities, version, logger)}}{
   This method is used to instantiate a \code{WMSLayer} object
 }
 \item{\code{getName()}}{
   Get layer name
 }
 \item{\code{getTitle()}}{
   Get layer title
 }
 \item{\code{getAbstract()}}{
   Get layer abstract
 }
 \item{\code{getKeywords()}}{
   Get layer keywords
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Layer}
\keyword{OGC}
\keyword{WMS}
