% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSBoundingBoxData.R
\docType{class}
\name{WPSBoundingBoxData}
\alias{WPSBoundingBoxData}
\title{WPSBoundingBoxData}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an OGC WPS BoundingBox data
}
\description{
WPSBoundingBoxData
}
\section{Methods}{

\describe{
 \item{\code{new(xml, owsVersion, serviceVersion, logger)}}{
   This method is used to instantiate an \code{WPSBoundingBoxData} object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Bounding}
\keyword{Box}
\keyword{OGC}
\keyword{WPS}
\keyword{data}
