% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSComplexData.R
\docType{class}
\name{WPSComplexData}
\alias{WPSComplexData}
\title{WPSComplexData}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS Complex Data
}
\description{
WPSComplexData
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value, schema, mimeType)}}{
   This method is used to instantiate a \code{WPSComplexData} object
 }
 \item{\code{decode(xml)}}{
   Decodes from XML
 }
 \item{\code{getFeatures()}}{
   Returns features associates with output, if the output si made of a GML feature collection
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ComplexData}
\keyword{OGC}
\keyword{WPS}
