% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSComplexInputDescription.R
\docType{class}
\name{WPSComplexInputDescription}
\alias{WPSComplexInputDescription}
\title{WPSComplexInputDescription}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process complex input description
}
\description{
WPSComplexInputDescription
}
\note{
Class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, version, logger)}}{
   This method is used to instantiate a \code{WPSComplexInputDescription} object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{WPS}
\keyword{complex}
\keyword{description}
\keyword{input}
