% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSDescriptionParameter.R
\docType{class}
\name{WPSDescriptionParameter}
\alias{WPSDescriptionParameter}
\title{WPSDescriptionParameter}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process input description parameter
}
\description{
WPSDescriptionParameter
}
\note{
Class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xml, version, logger)}}{
   This method is used to instantiate a \code{WPSDescriptionParameter} object
 }
 \item{\code{getFormats()}}{
   Get formats
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{WPS}
\keyword{description}
\keyword{input}
\keyword{parameter}
