% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSExecuteResponse.R
\docType{class}
\name{WPSExecuteResponse}
\alias{WPSExecuteResponse}
\title{WPSExecuteResponse}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS Execute response
}
\description{
WPSExecuteResponse
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj)}}{
   This method is used to instantiate a WPSExecuteResponse object
 }
 \item{\code{getProcess()}}{
   Get the process
 }
 \item{\code{getStatus()}}{
   Get the status
 }
 \item{\code{getProcessOutputs()}}{
   Get the process output(s)
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ExecuteResponse}
\keyword{OGC}
\keyword{WPS}
