% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSFormat.R
\docType{class}
\name{WPSFormat}
\alias{WPSFormat}
\title{WPSFormat}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process input description
}
\description{
WPSFormat

WPSFormat
}
\note{
Class used internally by \pkg{ows4R}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{WPS}
\keyword{description}
\keyword{input}
\section{Super class}{
\code{\link[ows4R:OGCAbstractObject]{ows4R::OGCAbstractObject}} -> \code{WPSFormat}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WPSFormat$new()}}
\item \href{#method-getMimeType}{\code{WPSFormat$getMimeType()}}
\item \href{#method-getEncoding}{\code{WPSFormat$getEncoding()}}
\item \href{#method-getSchema}{\code{WPSFormat$getSchema()}}
\item \href{#method-setIsDefault}{\code{WPSFormat$setIsDefault()}}
\item \href{#method-isDefault}{\code{WPSFormat$isDefault()}}
\item \href{#method-clone}{\code{WPSFormat$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="encode">}\href{../../ows4R/html/OGCAbstractObject.html#method-encode}{\code{ows4R::OGCAbstractObject$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="ERROR">}\href{../../ows4R/html/OGCAbstractObject.html#method-ERROR}{\code{ows4R::OGCAbstractObject$ERROR()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClass">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClass}{\code{ows4R::OGCAbstractObject$getClass()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getClassName">}\href{../../ows4R/html/OGCAbstractObject.html#method-getClassName}{\code{ows4R::OGCAbstractObject$getClassName()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="getNamespaceDefinition">}\href{../../ows4R/html/OGCAbstractObject.html#method-getNamespaceDefinition}{\code{ows4R::OGCAbstractObject$getNamespaceDefinition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="INFO">}\href{../../ows4R/html/OGCAbstractObject.html#method-INFO}{\code{ows4R::OGCAbstractObject$INFO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="isFieldInheritedFrom">}\href{../../ows4R/html/OGCAbstractObject.html#method-isFieldInheritedFrom}{\code{ows4R::OGCAbstractObject$isFieldInheritedFrom()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="logger">}\href{../../ows4R/html/OGCAbstractObject.html#method-logger}{\code{ows4R::OGCAbstractObject$logger()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ows4R" data-topic="OGCAbstractObject" data-id="WARN">}\href{../../ows4R/html/OGCAbstractObject.html#method-WARN}{\code{ows4R::OGCAbstractObject$WARN()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{WPSFormat}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSFormat$new(xml = NULL, version, logger = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{an object of class \link{XMLInternalNode-class} to initialize from XML}

\item{\code{version}}{WPS service version}

\item{\code{logger}}{logger}

\item{\code{...}}{any additional parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMimeType"></a>}}
\if{latex}{\out{\hypertarget{method-getMimeType}{}}}
\subsection{Method \code{getMimeType()}}{
Get mime type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSFormat$getMimeType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEncoding"></a>}}
\if{latex}{\out{\hypertarget{method-getEncoding}{}}}
\subsection{Method \code{getEncoding()}}{
Get encoding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSFormat$getEncoding()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSchema"></a>}}
\if{latex}{\out{\hypertarget{method-getSchema}{}}}
\subsection{Method \code{getSchema()}}{
Get schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSFormat$getSchema()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setIsDefault"></a>}}
\if{latex}{\out{\hypertarget{method-setIsDefault}{}}}
\subsection{Method \code{setIsDefault()}}{
set is default
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSFormat$setIsDefault(default)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{default}}{object of class \code{logical}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isDefault"></a>}}
\if{latex}{\out{\hypertarget{method-isDefault}{}}}
\subsection{Method \code{isDefault()}}{
is default
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSFormat$isDefault()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if default, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WPSFormat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
