% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSInput.R
\docType{class}
\name{WPSInput}
\alias{WPSInput}
\title{WPSInput}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS Input
}
\description{
WPSInput
}
\section{Methods}{

\describe{
 \item{\code{new(xml, identifier, data, serviceVersion)}}{
   This method is used to instantiate a WPSInput object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Input}
\keyword{OGC}
\keyword{WPS}
