% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSInputDescription.R
\docType{class}
\name{WPSInputDescription}
\alias{WPSInputDescription}
\title{WPSInputDescription}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process input description
}
\description{
WPSInputDescription
}
\note{
Class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xml, capabilities, version, logger)}}{
   This method is used to instantiate a \code{WPSInputDescription} object
 }
 \item{\code{getMinOccurs()}}{
   Get input min occurs
 }
 \item{\code{getMaxOccurs()}}{
   Get input max occurs
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{WPS}
\keyword{description}
\keyword{input}
