% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSLiteralData.R
\docType{class}
\name{WPSLiteralData}
\alias{WPSLiteralData}
\title{WPSLiteralData}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS Literal Data
}
\description{
WPSLiteralData
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, value)}}{
   This method is used to instantiate a \code{WPSLiteralData} object
 }
 \item{\code{decode()}}{
   Decodes WPS input from XML
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{LiteralData}
\keyword{OGC}
\keyword{WPS}
