% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSLiteralInputDescription.R
\docType{class}
\name{WPSLiteralInputDescription}
\alias{WPSLiteralInputDescription}
\title{WPSLiteralInputDescription}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS process input description
}
\description{
WPSLiteralInputDescription
}
\note{
Class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, version, logger)}}{
   This method is used to instantiate a \code{WPSLiteralInputDescription} object
 }
 \item{\code{getDataType()}}{
   Get data type
 }
 \item{\code{getDefaultValue()}}{
   Get default value
 }
 \item{\code{getAllowedValues()}}{
   Get allowed values
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Process}
\keyword{WPS}
\keyword{description}
\keyword{input}
