% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSOutput.R
\docType{class}
\name{WPSOutput}
\alias{WPSOutput}
\title{WPSOutput}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a WPS Input
}
\description{
WPSOutput
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, identifier, title, data)}}{
   This method is used to instantiate a \code{WPSOutput} object
 }
 \item{\code{decode()}}{
   Decodes WPS output from XML
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Input}
\keyword{OGC}
\keyword{WPS}
