% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPSParameter.R
\docType{class}
\name{WPSParameter}
\alias{WPSParameter}
\title{WPSParameter}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} modelling a WPS parameter
}
\description{
WPSParameter
}
\note{
Abstract class used internally by \pkg{ows4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xmlObj, version, logger)}}{
   This method is used to instantiate a \code{WPSParameter} object
 }
 \item{\code{getIdentifier()}}{
   Get input identifier
 }
 \item{\code{getTitle()}}{
   Get input title
 }
 \item{\code{getAbstract()}}{
   Get input abstract
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{OGC}
\keyword{Parameter}
\keyword{WPS}
